### 2023-01-18 #5085 @editorialbot
[PRE REVIEW]: HDT: A Rust library for the Header Data Triples binary RDF compression format
**Submitting author:** @KonradHoeffner (<a href="http://orcid.org/0000-0001-7358-3217">Konrad Höffner</a>)
**Repository:** https://github.com/konradhoeffner/hdt
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.0.11
