### 2023-01-18 #5085 @danielskatz
👋 @KonradHoeffner - thanks for your submission to JOSS. The next step is for me to find an editor for it.

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
