+++ src/hdt_graph.rs
@@ -2,243 +2,202 @@
 #[cfg(feature = "sophia")]
 use crate::four_sect_dict::IdKind;
 use crate::hdt::Hdt;
-use crate::predicate_object_iter::PredicateObjectIter;
-use crate::triples::{BitmapIter, TripleId};
+use crate::triples::{Id, ObjectIter, PredicateIter, PredicateObjectIter, SubjectIter, TripleId};
 use log::debug;
-use log::error;
-use sophia::graph::*;
-use sophia::term::iri::Iri;
-use sophia::term::literal::Literal;
-use sophia::term::*;
-use sophia::triple::stream::*;
-use sophia::triple::streaming_mode::*;
-use std::convert::From;
+
+use sophia::api::graph::{GTripleSource, Graph};
+//use sophia::api::term;
+use mownstr::MownStr;
+
+use sophia::api::term::FromTerm;
+use sophia::api::term::{matcher::TermMatcher, BnodeId, IriRef, LanguageTag, SimpleTerm, Term};
 use std::convert::Infallible;
-use std::hash::Hash;
-use std::marker::PhantomData;
-
-/// Type of String pointer to return, predetermined by Sophia.
-pub trait Data:
-    AsRef<str> + Clone + Eq + From<std::boxed::Box<str>> + From<String> + From<&'static str> + Hash
-{
-}
-impl<T> Data for T where
-    T: AsRef<str> + Clone + Eq + From<std::boxed::Box<str>> + From<String> + From<&'static str> + Hash
-{
-}
-
-/// Our own requirements on top of the Sophia trait.
-pub trait DisplayData: Data + std::fmt::Debug + std::fmt::Display {}
-impl<T> DisplayData for T where T: Data + std::fmt::Debug + std::fmt::Display {}
+use std::io::{self, Error, ErrorKind};
+use std::iter;
 
 /// Adapter to use HDT as a Sophia graph.
-pub struct HdtGraph<T: DisplayData> {
-    hdt: Hdt<T>,
-    phantom: PhantomData<T>,
+pub struct HdtGraph {
+    hdt: Hdt,
 }
 
-impl<T: DisplayData> HdtGraph<T> {
+impl HdtGraph {
     /// Wrapper around Hdt.
-    pub const fn new(hdt: Hdt<T>) -> Self {
-        HdtGraph { hdt, phantom: PhantomData {} }
+    pub const fn new(hdt: Hdt) -> Self {
+        HdtGraph { hdt }
     }
     /// Size in bytes on the heap.
     pub fn size_in_bytes(&self) -> usize {
         self.hdt.size_in_bytes()
     }
-}
 
-const XSD_STRING: &str = "http://www.w3.org/2001/XMLSchema#string";
+    fn id_term(&self, id: Id, kind: &'static IdKind) -> SimpleTerm<'static> {
+        IriRef::new_unchecked(MownStr::from(self.hdt.dict.id_to_string(id, kind).unwrap())).into_term()
+    }
+}
 
 /// Create the correct Sophia term for a given resource string.
 /// Slow, use the appropriate method if you know which type (Literal, URI, or blank node) the string has.
-fn auto_term<T: DisplayData>(s: &str) -> Result<Term<T>, TermError> {
+fn auto_term(s: MownStr) -> io::Result<SimpleTerm> {
     match s.chars().next() {
-        None => Err(TermError::InvalidIri(String::new())),
+        None => Err(Error::new(ErrorKind::InvalidData, "empty input")),
         Some('"') => match s.rfind('"') {
-            None => Err(TermError::UnsupportedDatatype(s.to_owned())),
+            None => Err(Error::new(
+                ErrorKind::InvalidData,
+                format!("missing right quotation mark in literal string {s}"),
+            )),
             Some(index) => {
[…]
