+++ src/hdt.rs
@@ -1,39 +1,28 @@
 use crate::containers::ControlInfo;
 use crate::four_sect_dict::{DictErr, IdKind};
 use crate::header::Header;
-use crate::predicate_object_iter::PredicateObjectIter;
-use crate::triples::{BitmapIter, TripleId, TriplesBitmap};
+use crate::triples::{ObjectIter, PredicateIter, PredicateObjectIter, SubjectIter, TripleId, TriplesBitmap};
 use crate::FourSectDict;
 use bytesize::ByteSize;
 use log::{debug, error};
+use mownstr::MownStr;
 use std::io;
-use std::marker::PhantomData;
+use std::iter;
 use thiserror::Error;
 
 /// In-memory representation of an RDF graph loaded from an HDT file.
 /// Allows queries by triple patterns.
 #[derive(Debug)]
-pub struct Hdt<T: Data> {
+pub struct Hdt {
     //global_ci: ControlInfo,
     //header: Header,
     /// in-memory representation of dictionary
     pub dict: FourSectDict,
     /// in-memory representation of triples
     pub triples: TriplesBitmap,
-    phantom: PhantomData<T>,
 }
 
-/// Type of String pointer to return, intended for Box, Rc and Arc.
-pub trait Data:
-    AsRef<str> + Clone + Eq + From<std::boxed::Box<str>> + From<String> + From<&'static str> + std::hash::Hash
-{
-}
-impl<T> Data for T where
-    T: AsRef<str> + Clone + Eq + From<std::boxed::Box<str>> + From<String> + From<&'static str> + std::hash::Hash
-{
-}
-
-type StringTriple<T> = (T, T, T);
+type StringTriple<'a> = (MownStr<'a>, MownStr<'a>, MownStr<'a>);
 
 /// The error type for the `translate_id` method.
 #[derive(Error, Debug)]
@@ -44,7 +33,7 @@ pub struct TranslateErr {
     t: TripleId,
 }
 
-impl<T: Data> Hdt<T> {
+impl Hdt {
     /// Creates an immutable HDT instance containing the dictionary and triples from the given reader.
     /// The reader must point to the beginning of the data of an HDT file as produced by hdt-cpp.
     /// FourSectionDictionary with DictionarySectionPlainFrontCoding and SPO order is the only supported implementation.
@@ -56,9 +45,9 @@ impl<T: Data> Hdt<T> {
         let mut dict = FourSectDict::read(&mut reader)?;
         let triples = TriplesBitmap::read_sect(&mut reader)?;
         dict.validate()?;
-        let hdt = Hdt { dict, triples, phantom: PhantomData {} };
+        let hdt = Hdt { dict, triples };
         debug!("HDT size in memory {}, details:", ByteSize(hdt.size_in_bytes() as u64));
-        //debug!("{hdt:#?}");
+        debug!("{hdt:#?}");
         Ok(hdt)
     }
 
@@ -67,8 +56,8 @@ impl<T: Data> Hdt<T> {
         self.dict.size_in_bytes() + self.triples.size_in_bytes()
     }
 
-    /// Don't use this for many triples with shared values as you won't benefit from deduplication if you are using a smartpointer like Rc<str> or Arc<str>.
-    fn translate_id(&self, t: TripleId) -> Result<StringTriple<T>, TranslateErr> {
+    /// Don't use this for many triples with shared values as you won't benefit from deduplication.
+    fn translate_id(&self, t: TripleId) -> Result<StringTriple, TranslateErr> {
         let s = self.dict.id_to_string(t.subject_id, &IdKind::Subject).map_err(|e| TranslateErr { e, t })?;
[…]
