### 2023-01-19 6d39e03
Start refactoring to align with Sophia 0.8.0-alpha.
