+++ src/hdt_graph.rs
@@ -6,12 +6,9 @@ use crate::predicate_object_iter::PredicateObjectIter;
 use crate::triples::{BitmapIter, TripleId};
 use log::debug;
 use log::error;
-use sophia::graph::*;
-use sophia::term::iri::Iri;
-use sophia::term::literal::Literal;
-use sophia::term::*;
-use sophia::triple::stream::*;
-use sophia::triple::streaming_mode::*;
+use sophia::api::graph::{GTripleSource, Graph};
+use sophia::api::term;
+use sophia::api::term::{SimpleTerm, SimpleTerm::*, Term, TermKind};
 use std::convert::From;
 use std::convert::Infallible;
 use std::hash::Hash;
@@ -26,6 +23,8 @@ impl<T> Data for T where
     T: AsRef<str> + Clone + Eq + From<std::boxed::Box<str>> + From<String> + From<&'static str> + Hash
 {
 }
+
+type TermError = 'static + std::error::Error;
 
 /// Our own requirements on top of the Sophia trait.
 pub trait DisplayData: Data + std::fmt::Debug + std::fmt::Display {}
@@ -52,7 +51,7 @@ const XSD_STRING: &str = "http://www.w3.org/2001/XMLSchema#string";
 
 /// Create the correct Sophia term for a given resource string.
 /// Slow, use the appropriate method if you know which type (Literal, URI, or blank node) the string has.
-fn auto_term<T: DisplayData>(s: &str) -> Result<Term<T>, TermError> {
+fn auto_term<T: DisplayData>(s: &str) -> Result<SimpleTerm<T>, TermError> {
     match s.chars().next() {
         None => Err(TermError::InvalidIri(String::new())),
         Some('"') => match s.rfind('"') {
@@ -64,7 +63,7 @@ fn auto_term<T: DisplayData>(s: &str) -> Result<Term<T>, TermError> {
                 if rest.is_empty() {
                     //let dt_string: sophia::term::iri::Iri<&str> = Iri::<&str>::new_unchecked(XSD_STRING);
                     let dt_string = Iri::<&str>::new_unchecked(XSD_STRING);
-                    return Ok(Term::<T>::from(Literal::new_dt(lex, dt_string)));
+                    return Ok(Term::<T>::from(LiteralDatatype(lex, dt_string)));
                 }
                 // either language tag or datatype
                 if let Some(tag_index) = rest.find('@') {
@@ -80,7 +79,7 @@ fn auto_term<T: DisplayData>(s: &str) -> Result<Term<T>, TermError> {
                     Some(dt) => {
                         let unquoted = dt[1..dt.len() - 1].to_owned().into_boxed_str();
                         let dt = Iri::<Box<str>>::new_unchecked(unquoted);
-                        Ok(Term::<T>::from(Literal::new_dt(lex, dt)))
+                        Ok(Term::<T>::from(LiteralDatatype(lex, dt)))
                     }
                     None => Err(TermError::UnsupportedDatatype(s.to_owned())),
                 }
@@ -101,11 +100,7 @@ fn triple_source<'s, T: DisplayData>(
                 debug_assert_ne!("", s.as_ref(), "triple_source subject is empty   ({s}, {p}, {o})");
                 debug_assert_ne!("", p.as_ref(), "triple_source predicate is empty ({s}, {p}, {o})");
                 debug_assert_ne!("", o.as_ref(), "triple_source object is empty    ({s}, {p}, {o})");
-                Ok(StreamedTriple::by_value([
-                    auto_term(s.as_ref())?,
-                    auto_term(p.as_ref())?,
-                    auto_term(o.as_ref())?,
-                ]))
[…]
