+++ Cargo.toml
@@ -17,7 +17,7 @@ langtag = "^0.3.2"
 ntriple = "^0.1.1"
 #rayon = "1.5.3"
 rsdict = "0.0.6"
-sophia = { version = "0.7", optional = true }
+sophia = { version="0.8.0-alpha", git="https://github.com/pchampin/sophia_rs", optional = true }
 sucds = "0.6.0"
 thiserror = "1.0.37"
 log = "0.4"
