### 2023-01-20 c78cf07
More refactoring to MownStr.
