+++ src/hdt_graph.rs
@@ -7,39 +7,22 @@ use crate::triples::{BitmapIter, TripleId};
 use log::debug;
 use log::error;
 use sophia::api::graph::{GTripleSource, Graph};
-use sophia::api::term;
+//use sophia::api::term;
+use mownstr::MownStr;
 use sophia::api::term::{SimpleTerm, SimpleTerm::*, Term, TermKind};
 use std::convert::From;
 use std::convert::Infallible;
-use std::hash::Hash;
-use std::marker::PhantomData;
-
-/// Type of String pointer to return, predetermined by Sophia.
-pub trait Data:
-    AsRef<str> + Clone + Eq + From<std::boxed::Box<str>> + From<String> + From<&'static str> + Hash
-{
-}
-impl<T> Data for T where
-    T: AsRef<str> + Clone + Eq + From<std::boxed::Box<str>> + From<String> + From<&'static str> + Hash
-{
-}
-
-type TermError = 'static + std::error::Error;
-
-/// Our own requirements on top of the Sophia trait.
-pub trait DisplayData: Data + std::fmt::Debug + std::fmt::Display {}
-impl<T> DisplayData for T where T: Data + std::fmt::Debug + std::fmt::Display {}
+use std::io::{Error, ErrorKind};
 
 /// Adapter to use HDT as a Sophia graph.
-pub struct HdtGraph<T: DisplayData> {
-    hdt: Hdt<T>,
-    phantom: PhantomData<T>,
+pub struct HdtGraph {
+    hdt: Hdt,
 }
 
-impl<T: DisplayData> HdtGraph<T> {
+impl HdtGraph {
     /// Wrapper around Hdt.
-    pub const fn new(hdt: Hdt<T>) -> Self {
-        HdtGraph { hdt, phantom: PhantomData {} }
+    pub const fn new(hdt: Hdt) -> Self {
+        HdtGraph { hdt }
     }
     /// Size in bytes on the heap.
     pub fn size_in_bytes(&self) -> usize {
@@ -47,53 +30,49 @@ impl<T: DisplayData> HdtGraph<T> {
     }
 }
 
-const XSD_STRING: &str = "http://www.w3.org/2001/XMLSchema#string";
-
 /// Create the correct Sophia term for a given resource string.
 /// Slow, use the appropriate method if you know which type (Literal, URI, or blank node) the string has.
-fn auto_term<T: DisplayData>(s: &str) -> Result<SimpleTerm<T>, TermError> {
+fn auto_term(s: &str) -> Result<SimpleTerm, Error> {
     match s.chars().next() {
-        None => Err(TermError::InvalidIri(String::new())),
+        None => Err(Error::new(ErrorKind::InvalidData, "empty input")),
         Some('"') => match s.rfind('"') {
-            None => Err(TermError::UnsupportedDatatype(s.to_owned())),
+            None => Err(Error::new(
+                ErrorKind::InvalidData,
+                format!("missing right quotation mark in literal string {}", s),
+            )),
             Some(index) => {
                 let lex = s[1..index].to_owned();
                 let rest = &s[index + 1..];
                 // literal with no language tag and no datatype
                 if rest.is_empty() {
-                    //let dt_string: sophia::term::iri::Iri<&str> = Iri::<&str>::new_unchecked(XSD_STRING);
-                    let dt_string = Iri::<&str>::new_unchecked(XSD_STRING);
-                    return Ok(Term::<T>::from(LiteralDatatype(lex, dt_string)));
+                    return Ok(SimpleTerm::from(lex));
                 }
                 // either language tag or datatype
                 if let Some(tag_index) = rest.find('@') {
-                    return Ok(Term::<T>::from(Literal::new_lang_unchecked(
-                        lex,
[…]
