+++ src/hdt.rs
@@ -6,34 +6,23 @@ use crate::triples::{BitmapIter, TripleId, TriplesBitmap};
 use crate::FourSectDict;
 use bytesize::ByteSize;
 use log::{debug, error};
+use mownstr::MownStr;
 use std::io;
-use std::marker::PhantomData;
 use thiserror::Error;
 
 /// In-memory representation of an RDF graph loaded from an HDT file.
 /// Allows queries by triple patterns.
 #[derive(Debug)]
-pub struct Hdt<T: Data> {
+pub struct Hdt {
     //global_ci: ControlInfo,
     //header: Header,
     /// in-memory representation of dictionary
     pub dict: FourSectDict,
     /// in-memory representation of triples
     pub triples: TriplesBitmap,
-    phantom: PhantomData<T>,
 }
 
-/// Type of String pointer to return, intended for Box, Rc and Arc.
-pub trait Data:
-    AsRef<str> + Clone + Eq + From<std::boxed::Box<str>> + From<String> + From<&'static str> + std::hash::Hash
-{
-}
-impl<T> Data for T where
-    T: AsRef<str> + Clone + Eq + From<std::boxed::Box<str>> + From<String> + From<&'static str> + std::hash::Hash
-{
-}
-
-type StringTriple<T> = (T, T, T);
+type StringTriple<'a> = (MownStr<'a>, MownStr<'a>, MownStr<'a>);
 
 /// The error type for the `translate_id` method.
 #[derive(Error, Debug)]
@@ -44,7 +33,7 @@ pub struct TranslateErr {
     t: TripleId,
 }
 
-impl<T: Data> Hdt<T> {
+impl Hdt {
     /// Creates an immutable HDT instance containing the dictionary and triples from the given reader.
     /// The reader must point to the beginning of the data of an HDT file as produced by hdt-cpp.
     /// FourSectionDictionary with DictionarySectionPlainFrontCoding and SPO order is the only supported implementation.
@@ -56,7 +45,7 @@ impl<T: Data> Hdt<T> {
         let mut dict = FourSectDict::read(&mut reader)?;
         let triples = TriplesBitmap::read_sect(&mut reader)?;
         dict.validate()?;
-        let hdt = Hdt { dict, triples, phantom: PhantomData {} };
+        let hdt = Hdt { dict, triples };
         debug!("HDT size in memory {}, details:", ByteSize(hdt.size_in_bytes() as u64));
         //debug!("{hdt:#?}");
         Ok(hdt)
@@ -68,7 +57,7 @@ impl<T: Data> Hdt<T> {
     }
 
     /// Don't use this for many triples with shared values as you won't benefit from deduplication if you are using a smartpointer like Rc<str> or Arc<str>.
-    fn translate_id(&self, t: TripleId) -> Result<StringTriple<T>, TranslateErr> {
+    fn translate_id(&self, t: TripleId) -> Result<StringTriple, TranslateErr> {
         let s = self.dict.id_to_string(t.subject_id, &IdKind::Subject).map_err(|e| TranslateErr { e, t })?;
         let p = self.dict.id_to_string(t.predicate_id, &IdKind::Predicate).map_err(|e| TranslateErr { e, t })?;
         let o = self.dict.id_to_string(t.object_id, &IdKind::Object).map_err(|e| TranslateErr { e, t })?;
@@ -79,7 +68,7 @@ impl<T: Data> Hdt<T> {
     /// Using this method with a filter can be inefficient for large graphs,
     /// because the strings are stored in compressed form and must be decompressed and allocated.
     /// Whenever possible, use [`Hdt::triples_with`] instead.
-    pub fn triples(&self) -> impl Iterator<Item = StringTriple<T>> + '_ {
[…]
