+++ Cargo.toml
@@ -21,6 +21,7 @@ sophia = { version="0.8.0-alpha", git="https://github.com/pchampin/sophia_rs", o
 sucds = "0.6.0"
 thiserror = "1.0.37"
 log = "0.4"
+mownstr = "0.1.3"
 
 [features]
 default = ["sophia"]
