### 2023-01-20 c2dac37
More refactoring.
