+++ Cargo.toml
@@ -21,7 +21,8 @@ sophia = { version="0.8.0-alpha", git="https://github.com/pchampin/sophia_rs", o
 sucds = "0.6.0"
 thiserror = "1.0.37"
 log = "0.4"
-mownstr = "0.1.3"
+#mownstr = "0.1.3"
+mownstr = { git="https://github.com/pchampin/mownstr", rev="c88467d" }
 
 [features]
 default = ["sophia"]
