+++ src/hdt_graph.rs
@@ -9,9 +9,10 @@ use log::error;
 use sophia::api::graph::{GTripleSource, Graph};
 //use sophia::api::term;
 use mownstr::MownStr;
+use sophia::api::source::IntoTripleSource;
 use sophia::api::term::{BnodeId, IriRef, LanguageTag, SimpleTerm, Term};
 use std::convert::Infallible;
-use std::io::{Error, ErrorKind};
+use std::io::{self, Error, ErrorKind};
 
 /// Adapter to use HDT as a Sophia graph.
 pub struct HdtGraph {
@@ -31,7 +32,7 @@ impl HdtGraph {
 
 /// Create the correct Sophia term for a given resource string.
 /// Slow, use the appropriate method if you know which type (Literal, URI, or blank node) the string has.
-fn auto_term(s: &str) -> Result<SimpleTerm, Error> {
+fn auto_term(s: &str) -> io::Result<SimpleTerm> {
     match s.chars().next() {
         None => Err(Error::new(ErrorKind::InvalidData, "empty input")),
         Some('"') => match s.rfind('"') {
@@ -76,13 +77,13 @@ fn triple_source<'a>(
 ) -> GTripleSource<'a, HdtGraph> {
     Box::new(
         triples
-            .map(|(s, p, o)| -> Result<_> {
+            .map(|(s, p, o)| -> Result<_, _> {
                 debug_assert_ne!("", s.as_ref(), "triple_source subject is empty   ({s}, {p}, {o})");
                 debug_assert_ne!("", p.as_ref(), "triple_source predicate is empty ({s}, {p}, {o})");
                 debug_assert_ne!("", o.as_ref(), "triple_source object is empty    ({s}, {p}, {o})");
                 Ok([auto_term(s.as_ref())?, auto_term(p.as_ref())?, auto_term(o.as_ref())?])
             })
-            .filter_map(|r| r.map_err(|e| error!("{e}")).ok())
+            .filter_map(|r| r.map_err(|e: Error| error!("{e}")).ok())
             .into_triple_source(),
     )
 }
@@ -90,9 +91,9 @@ fn triple_source<'a>(
 // Sophia doesn't include the _: prefix for blank node strings but HDT expects it
 // not needed for property terms, as they can't be blank nodes
 fn term_string(t: &SimpleTerm) -> String {
-    match t() {
-        SimpleTerm::BlankNode(_) => "_:".to_owned() + &t.value(),
-        SimpleTerm::Iri(_) => t.value().to_string(),
+    match t {
+        SimpleTerm::BlankNode(b) => "_:".to_owned() + &b.as_str(),
+        SimpleTerm::Iri(i) => i.as_str().to_owned(),
         SimpleTerm::LiteralLanguage(_, _) => {
             let value = format!("\"{}\"", t.value());
             if let Some(lang) = t.language() {
@@ -126,22 +127,26 @@ impl Graph for HdtGraph {
 }
 
 impl HdtGraph {
-    fn triples_with_s<'s, TS: Term + ?Sized>(&'s self, s: &'s TS) -> GTripleSource<'s, Self> {
+    fn triples_with_s<'s>(&'s self, s: &SimpleTerm) -> GTripleSource<'s, Self> {
         let ss = term_string(s);
         let sid = self.hdt.dict.string_to_id(&ss, &IdKind::Subject);
         if sid == 0 {
             return Box::new(std::iter::empty());
         }
         // very inefficient conversion between TS and Term, TODO improve
-        let s = auto_term(&ss).unwrap();
+        //let s = auto_term(&ss).unwrap();
         Box::new(
             BitmapIter::with_pattern(&self.hdt.triples, &TripleId::new(sid, 0, 0))
[…]
