### 2023-01-22 213aed0
More refactoring.
