### 2023-01-22 b7af814
More refactoring.
