### 2023-01-23 7d61e35
Sort by cached key when generating wavelet matrix. Release 0.0.12.
