+++ README.md
@@ -29,7 +29,7 @@ If you need any of the those features, consider using a SPARQL endpoint instead.
 
 ```toml
 [dependencies]
-hdt = "0.0.11"
+hdt = "0.0.12"
 ```
 
 ```rust
@@ -61,5 +61,5 @@ If you don't want to pull in the Sophia dependency, you can exclude the adapter:
 
 ```toml
 [dependencies]
-hdt = { version = "0.0.11", default-features = false }
+hdt = { version = "0.0.12", default-features = false }
 ```
