+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "hdt"
-version = "0.0.11"
+version = "0.0.12"
 repository = "https://github.com/konradhoeffner/hdt"
 authors = ["Tim Baccaert <tbaccaer@vub.be>", "Konrad Höffner"]
 license = "MIT"
