+++ paper.bib
@@ -17,6 +17,7 @@ journal={Web Semantics: Science, Services and Agents on the World Wide Web},
 volume={19},
 pages={22–41},
 year={2013},
+doi={10.2139/ssrn.3198999},
 publisher={Elsevier}
 }
 
@@ -26,6 +27,7 @@ title = {Exchange and Consumption of Huge {RDF} Data},
 year = {2012},
 booktitle={The Semantic Web: Research and Applications},
 pages = {437-452},
+doi = {10.1007/978-3-642-30284-8_36},
 publisher={Springer}
 }
 
@@ -102,6 +104,7 @@ url = {https://www2020devtrack.github.io/site/schedule}
   number={1},
   pages={95--104},
   year={2016},
+  doi={10.3233/sw-150172},
   publisher={IOS Press}
 }
 
@@ -114,6 +117,6 @@ url = {https://www2020devtrack.github.io/site/schedule}
   publisher="Springer Berlin Heidelberg",
   address="Berlin, Heidelberg",
   pages="277--293",
-  abstract="Hundreds of public SPARQL endpoints have been deployed on the Web, forming a novel decentralised infrastructure for querying billions of structured facts from a variety of sources on a plethora of topics. But is this infrastructure mature enough to support applications? For 427 public SPARQL endpoints registered on the DataHub, we conduct various experiments to test their maturity. Regarding discoverability, we find that only one-third of endpoints make descriptive meta-data available, making it difficult to locate or learn about their content and capabilities. Regarding interoperability, we find patchy support for established SPARQL features like ORDER BY as well as (understandably) for new SPARQL 1.1 features. Regarding efficiency, we show that the performance of endpoints for generic queries can vary by up to 3--4 orders of magnitude. Regarding availability, based on a 27-month long monitoring experiment, we show that only 32.2{\%} of public endpoints can be expected to have (monthly) ``two-nines'' uptimes of 99--100{\%}.",
+  doi="10.1007/978-3-642-41338-4_18",
   isbn="978-3-642-41338-4"
 }
