### 2023-01-26 01f161e
Refactor iterators into the triple module.
