+++ src/triples.rs
@@ -1,6 +1,4 @@
 use crate::containers::{AdjList, Bitmap, Sequence};
-use crate::object_iter::ObjectIter;
-use crate::predicate_iter::PredicateIter;
 use crate::{ControlInfo, IdKind};
 use bytesize::ByteSize;
 use log::{debug, error};
@@ -11,6 +9,15 @@ use std::fmt;
 use std::io;
 use std::io::BufRead;
 use sucds::{CompactVector, Searial, WaveletMatrix, WaveletMatrixBuilder};
+
+mod subject_iter;
+pub use subject_iter::SubjectIter;
+mod predicate_iter;
+pub use predicate_iter::PredicateIter;
+mod predicate_object_iter;
+pub use predicate_object_iter::PredicateObjectIter;
+mod object_iter;
+pub use object_iter::ObjectIter;
 
 /// Order of the triple sections.
 /// Only SPO is tested, others probably don't work correctly.
@@ -126,7 +133,7 @@ impl TriplesBitmap {
     /// Iterator over all triples with a given ID in the specified position (subject, predicate or object).
     pub fn triples_with_id(&self, id: usize, id_kind: &IdKind) -> Box<dyn Iterator<Item = TripleId> + '_> {
         match id_kind {
-            IdKind::Subject => Box::new(BitmapIter::with_s(self, id)),
+            IdKind::Subject => Box::new(SubjectIter::with_s(self, id)),
             IdKind::Predicate => Box::new(PredicateIter::new(self, id)),
             IdKind::Object => Box::new(ObjectIter::new(self, id)),
         }
@@ -286,140 +293,10 @@ impl TriplesBitmap {
 
 impl<'a> IntoIterator for &'a TriplesBitmap {
     type Item = TripleId;
-    type IntoIter = BitmapIter<'a>;
+    type IntoIter = SubjectIter<'a>;
 
     fn into_iter(self) -> Self::IntoIter {
-        BitmapIter::new(self)
-    }
-}
-
-/// Iterator over triples fitting an SPO, SP? S?? or ??? triple pattern.
-//#[derive(Debug)]
-pub struct BitmapIter<'a> {
-    // triples data
-    triples: &'a TriplesBitmap,
-    // x-coordinate identifier
-    x: Id,
-    // current position
-    pos_y: usize,
-    pos_z: usize,
-    max_y: usize,
-    max_z: usize,
-}
-
-impl<'a> BitmapIter<'a> {
-    /// Create an iterator over all triples.
-    pub const fn new(triples: &'a TriplesBitmap) -> Self {
-        BitmapIter {
-            triples,
-            x: 1, // was 0 in the old code but it should start at 1
-            pos_y: 0,
-            pos_z: 0,
-            max_y: triples.wavelet_y.len(), // exclusive
-            max_z: triples.adjlist_z.len(), // exclusive
-        }
-    }
-
-    /// Use when no results are found.
-    pub const fn empty(triples: &'a TriplesBitmap) -> Self {
-        BitmapIter { triples, x: 1, pos_y: 0, pos_z: 0, max_y: 0, max_z: 0 }
-    }
-
-    /// see <https://github.com/rdfhdt/hdt-cpp/blob/develop/libhdt/src/triples/BitmapTriplesIterators.cpp>
-    pub fn with_s(triples: &'a TriplesBitmap, subject_id: Id) -> Self {
-        let min_y = triples.find_y(subject_id - 1);
-        let min_z = triples.adjlist_z.find(min_y as Id);
-        let max_y = triples.find_y(subject_id);
-        let max_z = triples.adjlist_z.find(max_y as Id);
-        BitmapIter { triples, x: subject_id, pos_y: min_y, pos_z: min_z, max_y, max_z }
[…]

[…]
