### 2023-01-26 076aca0
Start implementing hdt::triples_with_pattern.
