### 2023-01-27 #5085 @osorensen
:wave: @Yuri-M-Dias @remram44 @lazear @AustinTSchaffer would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
