### 2023-01-27 c5d3dff
Finish implementing triples_with_pattern.
