+++ src/hdt.rs
@@ -1,7 +1,7 @@
 use crate::containers::ControlInfo;
 use crate::four_sect_dict::{DictErr, IdKind};
 use crate::header::Header;
-use crate::triples::{PredicateObjectIter, SubjectIter, TripleId, TriplesBitmap};
+use crate::triples::{ObjectIter, PredicateIter, PredicateObjectIter, SubjectIter, TripleId, TriplesBitmap};
 use crate::FourSectDict;
 use bytesize::ByteSize;
 use log::{debug, error};
@@ -72,28 +72,6 @@ impl Hdt {
         self.triples.into_iter().map(|id| self.translate_id(id).unwrap())
     }
 
-    /// An iterator visiting all triples for a given triple pattern with exactly two variables, i.e. S??, ?P? or ??O.
-    /// Returns translated triples as strings.
-    /// If the subject is given, you can also use [`SubjectIter::with_pattern`] with a `TripleId` where property and object are 0.
-    /// Much more effient than filtering the result of [`Hdt::triples`].
-    /// If you want to query triple patterns with only one variable, use `triples_with_sp` etc. instead.
-    pub fn triples_with(&self, s: &str, kind: &'static IdKind) -> Box<dyn Iterator<Item = StringTriple> + '_> {
-        debug_assert_ne!("", s);
-        let id = self.dict.string_to_id(s, kind);
-        if id == 0 {
-            return Box::new(iter::empty());
-        }
-        let owned = s.to_owned();
-        Box::new(
-            self.triples
-                .triples_with_id(id, kind)
-                .map(move |tid| self.translate_id(tid))
-                .filter_map(move |r| r.map_err(|e| error!("Error on triple with {kind:?} {owned}: {e}")).ok()),
-        )
-    }
-
-    // TODO extract common code out of triples_with_...
-
     /// Get all triples with the given subject and property.
     pub fn objects_with_sp(&self, s: &str, p: &str) -> Box<dyn Iterator<Item = String> + '_> {
         let sid = self.dict.string_to_id(s, &IdKind::Subject);
@@ -110,39 +88,6 @@ impl Hdt {
                     r.map_err(|e| error!("Error on triple with subject {s_owned} and property {p_owned}: {e}"))
                         .ok()
                 }),
-        )
-    }
-
-    /// Get all triples with the given subject and property.
-    /// Inefficient with large results because the subject and property strings are duplicated for each triple.
-    /// Consider using [Self::objects_with_sp()] instead.
-    pub fn triples_with_sp<'a>(&'a self, s: &'a str, p: &'a str) -> impl Iterator<Item = StringTriple> {
-        let st = MownStr::from(s);
-        let pt = MownStr::from(p);
-        self.objects_with_sp(s, p).map(move |o| (st.clone(), pt.clone(), MownStr::from(o)))
-    }
-
-    /// Get all triples with the given subject and object.
-    /// The current implementation queries all triple IDs for the given subject and filters them for the given object.
-    /// This method is faster then filtering on translated strings but can be further optimized by creating a special SO iterator.
-    pub fn triples_with_so<'a>(&'a self, s: &'a str, o: &'a str) -> Box<dyn Iterator<Item = StringTriple> + '_> {
-        let sid = self.dict.string_to_id(s, &IdKind::Subject);
-        let oid = self.dict.string_to_id(o, &IdKind::Object);
-        if sid == 0 || oid == 0 {
-            return Box::new(iter::empty());
-        }
-        let st = MownStr::from(s);
-        let ot = MownStr::from(o);
-        Box::new(
-            self.triples
-                .triples_with_id(sid, &IdKind::Subject)
[…]
