### 2023-01-28 d2d377d
CI: Add GitHub Actions configuration
