### 2023-01-28 cbea6ca
CI: Only test on nightly Rust
