### 2023-01-29 ae3d61a
More refactoring.
