### 2023-01-29 46a14b2
More refactoring, now in hdt_graph.rs.
