+++ src/hdt_graph.rs
@@ -2,8 +2,7 @@
 #[cfg(feature = "sophia")]
 use crate::four_sect_dict::IdKind;
 use crate::hdt::Hdt;
-use crate::predicate_object_iter::PredicateObjectIter;
-use crate::triples::{BitmapIter, TripleId};
+use crate::triples::{ObjectIter, PredicateIter, PredicateObjectIter, SubjectIter, TripleId};
 use log::debug;
 use log::error;
 use sophia::api::graph::{GTripleSource, Graph};
@@ -14,6 +13,7 @@ use sophia::api::source::{IntoTripleSource, TripleSource};
 use sophia::api::term::{matcher::TermMatcher, BnodeId, IriRef, LanguageTag, SimpleTerm, Term};
 use std::convert::Infallible;
 use std::io::{self, Error, ErrorKind};
+use std::iter;
 
 /// Adapter to use HDT as a Sophia graph.
 pub struct HdtGraph {
@@ -116,24 +116,151 @@ impl Graph for HdtGraph {
         debug!("Iterating through ALL triples in the HDT Graph. This can be inefficient for large graphs.");
         Box::new(self.hdt.triples().map(move |(s, p, o)| {
             Ok([
-                SimpleTerm::Iri(IriRef::new_unchecked("".into())),
-                SimpleTerm::Iri(IriRef::new_unchecked("".into())),
+                SimpleTerm::Iri(IriRef::new_unchecked(s.into())),
+                SimpleTerm::Iri(IriRef::new_unchecked(p.into())),
                 SimpleTerm::Iri(IriRef::new_unchecked("".into())),
             ])
         }))
     }
 
+    /// Only supports constant and "any" matchers.
+    /// Non-constant matchers are supposed to be "any" matchers.
     fn triples_matching<'s, S, P, O>(&'s self, sm: S, pm: P, om: O) -> GTripleSource<'s, Self>
     where
         S: TermMatcher + 's,
         P: TermMatcher + 's,
         O: TermMatcher + 's,
     {
-        Box::new(
-            self.triples().filter(|r| r.is_ok()).filter(move |t| {
-                t.as_ref().unwrap().matched_by(sm.matcher_ref(), pm.matcher_ref(), om.matcher_ref())
-            }),
-        )
+        let xso = sm.constant().map(|s| {
+            //let simple = s.as_simple();
+            //let simple = SimpleTerm::from_term_ref(s);
+            let simple = SimpleTerm::from(s.as_simple().to_owned());
+            let id = self.hdt.dict.string_to_id(&term_string(&simple), &IdKind::Subject);
+            (simple, id)
+        });
+        let xpo = pm.constant().map(|p| {
+            let simple = p.as_simple();
+            let id = self.hdt.dict.string_to_id(&term_string(&simple), &IdKind::Predicate);
+            (simple, id)
+        });
+        let xoo = om.constant().map(|o| {
+            let simple = o.as_simple();
+            let id = self.hdt.dict.string_to_id(&term_string(&simple), &IdKind::Object);
+            (simple, id)
+        });
+        if [&xso, &xpo, &xoo].into_iter().flatten().any(|x| x.1 == 0) {
+            // at least one term does not exist in the graph
+            return Box::new(iter::empty());
+        }
+        //return Box::new(iter::empty());
+        if [&xso, &xpo, &xoo].into_iter().flatten().any(|x| x.1 == 0) {
+            // at least one term does not exist in the graph
+            return Box::new(iter::empty());
+        }
+        // TODO: improve error handling
+        match (xso, xpo, xoo) {
+            /*
+            (Some(s), Some(p), Some(o)) => {
+                if SubjectIter::with_pattern(&self.hdt.triples, &TripleId::new(s.1, p.1, o.1)).next().is_some() {
[…]
