### 2023-01-29 #5114 @remram44
Hi @KonradHoeffner, before I dig into the software: the paper looks good but I'm wondering if you should include a link to the [benchmarking repository](https://github.com/KonradHoeffner/sophia_benchmark)? It is quite a comprehensive benchmark and it looks like it took some effort to perform.
