### 2023-01-30 #5114 @KonradHoeffner
> Hi @KonradHoeffner, before I dig into the software: the paper looks good but I'm wondering if you should include a link to the [benchmarking repository](https://github.com/KonradHoeffner/sophia_benchmark)? It is quite a comprehensive benchmark and it looks like it took some effort to perform.

Hi, thanks for the suggestion! I have added a link to the HDT benchmark repository to the paper. As almost all of the benchmarking code comes from [Pierre-Antoine Champin](https://github.com/pchampin), I also added a link to his base repository that the HDT benchmark is a fork of.
In the README.md of HDT, there is also a "benchmark" button at the top, which links to the Juypter Lab notebook.
