### 2023-01-30 552b4cd
Add setup section and document nightly toolchain. Resolve #24.
