### 2023-01-30 7b7dc99
More refactoring. It compiles now. Slowly fixing tests and adapting code.
