+++ src/lib.rs
@@ -20,7 +20,7 @@
 //! let file = std::fs::File::open("example.hdt").expect("error opening file");
 //! let hdt = Hdt::new(std::io::BufReader::new(file)).expect("error loading HDT");
 //! // query
-//! let majors = hdt.triples_with_sp("http://dbpedia.org/resource/Leipzig", "http://dbpedia.org/ontology/major");
+//! let majors = hdt.triples_with_pattern(Some("http://dbpedia.org/resource/Leipzig"), Some("http://dbpedia.org/ontology/major"),None);
 //! println!("{:?}", majors.collect::<Vec<_>>());
 //! ```
 //!
@@ -28,14 +28,15 @@
 //!
 //! ```no_run
 //! use hdt::{Hdt,HdtGraph};
-//! use sophia::term::BoxTerm;
-//! use sophia::graph::Graph;
+//! use sophia::api::graph::Graph;
+//! use sophia::api::term::{IriRef, SimpleTerm, matcher::Any};
+//! use mownstr::MownStr;
 //! let file = std::fs::File::open("dbpedia.hdt").expect("error opening file");
 //! let hdt = Hdt::new(std::io::BufReader::new(file)).expect("error loading HDT");
 //! let graph = HdtGraph::new(hdt);
-//! let s = BoxTerm::new_iri_unchecked("http://dbpedia.org/resource/Leipzig");
-//! let p = BoxTerm::new_iri_unchecked("http://dbpedia.org/ontology/major");
-//! let majors = graph.triples_with_sp(&s,&p);
+//! let s = SimpleTerm::Iri(IriRef::new_unchecked(MownStr::from_str("http://dbpedia.org/resource/Leipzig")));
+//! let p = SimpleTerm::Iri(IriRef::new_unchecked(MownStr::from_str("http://dbpedia.org/ontology/major")));
+//! let majors = graph.triples_matching(Some(s),Some(p),Any);
 //! ```
 //!
 //! # Optional features
