+++ src/hdt_graph.rs
@@ -2,7 +2,7 @@
 #[cfg(feature = "sophia")]
 use crate::four_sect_dict::IdKind;
 use crate::hdt::Hdt;
-use crate::triples::{ObjectIter, PredicateIter, PredicateObjectIter, SubjectIter, TripleId};
+use crate::triples::{Id, ObjectIter, PredicateIter, PredicateObjectIter, SubjectIter, TripleId};
 use log::debug;
 use log::error;
 use sophia::api::graph::{GTripleSource, Graph};
@@ -10,6 +10,7 @@ use sophia::api::graph::{GTripleSource, Graph};
 use mownstr::MownStr;
 use sophia::api::prelude::Triple;
 use sophia::api::source::{IntoTripleSource, TripleSource};
+use sophia::api::term::FromTerm;
 use sophia::api::term::{matcher::TermMatcher, BnodeId, IriRef, LanguageTag, SimpleTerm, Term};
 use std::convert::Infallible;
 use std::io::{self, Error, ErrorKind};
@@ -29,11 +30,15 @@ impl HdtGraph {
     pub fn size_in_bytes(&self) -> usize {
         self.hdt.size_in_bytes()
     }
+
+    fn term(&self, id: Id, kind: &'static IdKind) -> SimpleTerm<'static> {
+        IriRef::new_unchecked(MownStr::from(self.hdt.dict.id_to_string(id, kind).unwrap())).into_term()
+    }
 }
 
 /// Create the correct Sophia term for a given resource string.
 /// Slow, use the appropriate method if you know which type (Literal, URI, or blank node) the string has.
-fn auto_term<'a>(s: &'a str) -> io::Result<SimpleTerm<'a>> {
+fn auto_term(s: MownStr) -> io::Result<SimpleTerm> {
     match s.chars().next() {
         None => Err(Error::new(ErrorKind::InvalidData, "empty input")),
         Some('"') => match s.rfind('"') {
@@ -52,7 +57,7 @@ fn auto_term<'a>(s: &'a str) -> io::Result<SimpleTerm<'a>> {
                 // either language tag or datatype
                 if let Some(tag_index) = rest.find('@') {
                     let tag = LanguageTag::new_unchecked(MownStr::from_str(&rest[tag_index + 1..]));
-                    return Ok(SimpleTerm::LiteralLanguage(lex, tag));
+                    return Ok(SimpleTerm::from_term(SimpleTerm::LiteralLanguage(lex, tag)));
                 }
                 // datatype
                 let mut dt_split = rest.split("^^");
@@ -61,35 +66,18 @@ fn auto_term<'a>(s: &'a str) -> io::Result<SimpleTerm<'a>> {
                     Some(dt) => {
                         let unquoted = &dt[1..dt.len() - 1];
                         let dt = IriRef::new_unchecked(MownStr::from_str(unquoted));
-                        Ok(SimpleTerm::LiteralDatatype(lex, dt))
+                        Ok(SimpleTerm::from_term(SimpleTerm::LiteralDatatype(lex, dt)))
                     }
                     None => Err(Error::new(ErrorKind::InvalidData, format!("empty datatype in {s}"))),
                 }
             }
         },
         Some('_') => Ok(BnodeId::new_unchecked(MownStr::from_str(&s[2..].to_owned())).into_term()),
-        _ => Ok(SimpleTerm::Iri(IriRef::new_unchecked(MownStr::from_str(s)))),
+        _ => Ok(SimpleTerm::Iri(IriRef::new_unchecked(s))),
     }
 }
-/*
-// transforms Hdt triples into a sophia TripleSource
-fn triple_source<'a>(
[…]
