+++ src/hdt_graph.rs
@@ -85,10 +85,16 @@ fn term_string(t: &SimpleTerm) -> String {
         SimpleTerm::BlankNode(b) => "_:".to_owned() + &b.as_str(),
         SimpleTerm::Iri(i) => i.as_str().to_owned(),
         SimpleTerm::LiteralLanguage(l, lang) => {
-            format!("{l}@{}", lang.as_str())
+            format!("\"{l}\"@{}", lang.as_str())
         }
         SimpleTerm::LiteralDatatype(l, dt) => {
-            format!("{l}@{}", dt.as_str())
+            let xsd_string: &str = "http://www.w3.org/2001/XMLSchema#string";
+            let dts = dt.as_str();
+            if dts == xsd_string {
+                format!("\"{l}\"")
+            } else {
+                format!("\"{l}\"^^<{}>", dts)
+            }
         }
         _ => {
             panic!("Variable term strings and RDF-star are not supported.");
@@ -134,11 +140,6 @@ impl Graph for HdtGraph {
             // at least one term does not exist in the graph
             return Box::new(iter::empty());
         }
-        //return Box::new(iter::empty());
-        if [&xso, &xpo, &xoo].into_iter().flatten().any(|x| x.1 == 0) {
-            // at least one term does not exist in the graph
-            return Box::new(iter::empty());
-        }
         // TODO: improve error handling
         match (xso, xpo, xoo) {
             (Some(s), Some(p), Some(o)) => {
@@ -153,10 +154,9 @@ impl Graph for HdtGraph {
                     Ok([
                         s.0.clone(),
                         p.0.clone(),
-                        IriRef::new_unchecked(MownStr::from(
-                            self.hdt.dict.id_to_string(t.object_id, &IdKind::Object).unwrap(),
-                        ))
-                        .into_term(),
+                        auto_term(self.hdt.dict.id_to_string(t.object_id, &IdKind::Object).unwrap().into())
+                            .unwrap()
+                            .into_term(),
                     ])
                 }))
             }
@@ -180,9 +180,10 @@ impl Graph for HdtGraph {
                             self.hdt.dict.id_to_string(t.predicate_id, &IdKind::Predicate).unwrap(),
                         ))
                         .into_term(),
-                        IriRef::new_unchecked(MownStr::from(
+                        auto_term(MownStr::from(
                             self.hdt.dict.id_to_string(t.object_id, &IdKind::Object).unwrap(),
                         ))
+                        .expect("auto term failed with object")
                         .into_term(),
                     ])
                 }))
@@ -214,10 +215,8 @@ impl Graph for HdtGraph {
             })),
             (None, None, Some(o)) => Box::new(ObjectIter::new(&self.hdt.triples, o.1).map(move |t| {
                 Ok([
-                    IriRef::new_unchecked(MownStr::from(
-                        self.hdt.dict.id_to_string(t.subject_id, &IdKind::Subject).unwrap(),
-                    ))
-                    .into_term(),
+                    auto_term(MownStr::from(self.hdt.dict.id_to_string(t.subject_id, &IdKind::Subject).unwrap()))
+                        .unwrap(),
                     IriRef::new_unchecked(MownStr::from(
                         self.hdt.dict.id_to_string(t.predicate_id, &IdKind::Predicate).unwrap(),
                     ))
@@ -237,6 +236,7 @@ mod tests {
     use sophia::api::term::matcher::Any;
     use std::fs::File;
     use std::rc::Rc;
+    use std::result::Result;
 
     #[test]
[…]
