+++ src/hdt_graph.rs
@@ -4,12 +4,11 @@ use crate::four_sect_dict::IdKind;
 use crate::hdt::Hdt;
 use crate::triples::{Id, ObjectIter, PredicateIter, PredicateObjectIter, SubjectIter, TripleId};
 use log::debug;
-use log::error;
+
 use sophia::api::graph::{GTripleSource, Graph};
 //use sophia::api::term;
 use mownstr::MownStr;
-use sophia::api::prelude::Triple;
-use sophia::api::source::{IntoTripleSource, TripleSource};
+
 use sophia::api::term::FromTerm;
 use sophia::api::term::{matcher::TermMatcher, BnodeId, IriRef, LanguageTag, SimpleTerm, Term};
 use std::convert::Infallible;
@@ -31,7 +30,7 @@ impl HdtGraph {
         self.hdt.size_in_bytes()
     }
 
-    fn term(&self, id: Id, kind: &'static IdKind) -> SimpleTerm<'static> {
+    fn id_term(&self, id: Id, kind: &'static IdKind) -> SimpleTerm<'static> {
         IriRef::new_unchecked(MownStr::from(self.hdt.dict.id_to_string(id, kind).unwrap())).into_term()
     }
 }
@@ -44,7 +43,7 @@ fn auto_term(s: MownStr) -> io::Result<SimpleTerm> {
         Some('"') => match s.rfind('"') {
             None => Err(Error::new(
                 ErrorKind::InvalidData,
-                format!("missing right quotation mark in literal string {}", s),
+                format!("missing right quotation mark in literal string {s}"),
             )),
             Some(index) => {
                 let lex = &s[1..index];
@@ -72,7 +71,7 @@ fn auto_term(s: MownStr) -> io::Result<SimpleTerm> {
                 }
             }
         },
-        Some('_') => Ok(BnodeId::new_unchecked(MownStr::from_str(&s[2..].to_owned())).into_term()),
+        Some('_') => Ok(BnodeId::new_unchecked(MownStr::from_str(&s[2..])).into_term()),
         _ => Ok(SimpleTerm::Iri(IriRef::new_unchecked(s))),
     }
 }
@@ -82,7 +81,7 @@ fn auto_term(s: MownStr) -> io::Result<SimpleTerm> {
 // not needed for property terms, as they can't be blank nodes
 fn term_string(t: &SimpleTerm) -> String {
     match t {
-        SimpleTerm::BlankNode(b) => "_:".to_owned() + &b.as_str(),
+        SimpleTerm::BlankNode(b) => "_:".to_owned() + b.as_str(),
         SimpleTerm::Iri(i) => i.as_str().to_owned(),
         SimpleTerm::LiteralLanguage(l, lang) => {
             format!("\"{l}\"@{}", lang.as_str())
@@ -93,7 +92,7 @@ fn term_string(t: &SimpleTerm) -> String {
             if dts == xsd_string {
                 format!("\"{l}\"")
             } else {
-                format!("\"{l}\"^^<{}>", dts)
+                format!("\"{l}\"^^<{dts}>")
             }
         }
         _ => {
@@ -109,7 +108,7 @@ impl Graph for HdtGraph {
     fn triples(&self) -> GTripleSource<Self> {
         debug!("Iterating through ALL triples in the HDT Graph. This can be inefficient for large graphs.");
         Box::new(self.hdt.triples().map(move |(s, p, o)| {
-            Ok([auto_term(s).unwrap(), SimpleTerm::Iri(IriRef::new_unchecked(p.into())), auto_term(o).unwrap()])
+            Ok([auto_term(s).unwrap(), SimpleTerm::Iri(IriRef::new_unchecked(p)), auto_term(o).unwrap()])
         }))
     }
 
[…]
