### 2023-02-03 81e4d8a
Test explicit vs implicit xsd:string.
