### 2023-02-06 6e9e3a8
Add predicate check to PredicateObjectIter.
