### 2023-02-06 d8aaa33
Optimize extract function.
