+++ benches/bench.rs
@@ -1,48 +1,57 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 use hdt::triples::*;
 use hdt::Hdt;
+use hdt::IdKind;
 use sophia::api::ns::rdf;
 use std::fs::File;
 
-fn bench_subject_iter(c: &mut Criterion) {
-    let dbo_person = "http://dbpedia.org/ontology/Person";
-    let dbr_vincent = "http://dbpedia.org/resource/Vincent_Descombes_Sevoie";
+fn bench_query(c: &mut Criterion) {
+    let vincent = "http://dbpedia.org/resource/Vincent_Descombes_Sevoie";
     let type_ = rdf::type_.to_string();
+    let person = "http://dbpedia.org/ontology/Person";
     let file = File::open("tests/resources/persondata_en.hdt").expect("error opening file");
     //let file = File::open("tests/resources/lscomplete2015.hdt").expect("error opening file");
     //let file = File::open("tests/resources/snikmeta.hdt").expect("error opening file");
     let hdt = Hdt::new(std::io::BufReader::new(file)).unwrap();
     let triples = &hdt.triples;
+    let vincent_id = hdt.dict.string_to_id(vincent, &IdKind::Subject);
+    let type_id = hdt.dict.string_to_id(&type_, &IdKind::Predicate);
+    let person_id = hdt.dict.string_to_id(person, &IdKind::Object);
+    let twp = |s, p, o| hdt.triples_with_pattern(s, p, o);
+
     // count to prevent optimizing away function call
-    /*
-    let mut group = c.benchmark_group("query triple IDs");
-    group.bench_function("S??", |b| {
-        b.iter(|| SubjectIter::with_pattern(&triples, &TripleId::new(1, 0, 0)).count())
+    //let mut group = c.benchmark_group("S??");
+    let mut group = c.benchmark_group("query");
+    group.bench_function("1 triple IDs (vincent, ?, ?)", |b| {
+        b.iter(|| SubjectIter::with_pattern(&triples, &TripleId::new(vincent_id, 0, 0)).count())
     });
-    group.sample_size(10);
-    group.bench_function("?P?", |b| {
-        b.iter(|| SubjectIter::with_pattern(&triples, &TripleId::new(0, 1, 0)).count())
-    });
-    group.bench_function("??O", |b| {
-        b.iter(|| SubjectIter::with_pattern(&triples, &TripleId::new(0, 0, 1)).count())
-    });
-    group.finish();
-    */
-    let mut group = c.benchmark_group("query string triples");
-    group.bench_function("(vincent, ?, ?)", |b| {
-        b.iter(|| hdt.triples_with_pattern(Some(dbr_vincent), None, None).count())
-    });
+    group.bench_function("2 str triples (vincent, ?, ?)", |b| b.iter(|| twp(Some(vincent), None, None).count()));
+    //group.finish();
+    //let mut group = c.benchmark_group("?P?");
     group.sample_size(10);
     group
-        .bench_function("(?, type, ?)", |b| b.iter(|| hdt.triples_with_pattern(None, Some(&type_), None).count()));
-    group.bench_function("(?, ?, dbo_person)", |b| {
-        b.iter(|| hdt.triples_with_pattern(None, None, Some(dbo_person)).count())
+        .bench_function("3 triple IDs (?, type, ?)", |b| b.iter(|| PredicateIter::new(&triples, type_id).count()));
+    group.bench_function("4 str triples (?, type, ?)", |b| b.iter(|| twp(None, Some(&type_), None).count()));
+    //group.finish();
+    //let mut group = c.benchmark_group("??O");
+    group.bench_function("5 triple IDs (?, ?, person)", |b| {
[…]
