### 2023-02-07 d2092f6
Create alternative predicate object iterator.
