### 2023-02-07 b2cdb4a
Optimize PredicateObjectIter.
