### 2023-02-08 6363410
Point OP-index to the P-layer instead of the O-layer. Reduces time for large ??O and ?PO queries by about 10%.
