+++ paper.bib
@@ -120,3 +120,23 @@ url = {https://www2020devtrack.github.io/site/schedule}
   doi="10.1007/978-3-642-41338-4_18",
   isbn="978-3-642-41338-4"
 }
+
+@inproceedings{serializingrdf,
+  title={Serializing {RDF} in compressed space},
+  author={Hern{\'a}ndez-Illera, Antonio and Mart{\'\i}nez-Prieto, Miguel A and Fern{\'a}ndez, Javier D},
+  booktitle={2015 Data Compression Conference},
+  pages={363--372},
+  year={2015},
+  organization={IEEE}
+}
+
+@article{ihdt,
+  title={{iHDT++}: improving {HDT} for {SPARQL} triple pattern resolution},
+  author={Hern{\'a}ndez-Illera, Antonio and Mart{\'\i}nez-Prieto, Miguel A and Fern{\'a}ndez, Javier D and Fari{\~n}a, Antonio},
+  journal={Journal of Intelligent \& Fuzzy Systems},
+  volume={39},
+  number={2},
+  pages={2249--2261},
+  year={2020},
+  publisher={IOS Press}
+}
