### 2023-02-08 #5114 @KonradHoeffner
Recreated the benchmarks with new optimizations of the library. Also Included hdt-cpp in the benchmarks and unpacked the DelayedString instances in the hdt-java benchmark to take the time needed for that into account.
