### 2023-02-11 ec77d82
remove unnecessary utf-8 decoding
