### 2023-02-11 08880f8
return Error if parsing format or prop_str fails
these branches are NOT unreachable;
for example, the file might be incomplete,
in which case format/prop_str might be empty or not finished by 0x00.
You probably don't want to panic in that case.
