### 2023-02-11 78ab0c9
compute CRC incrementally in control_info.rs
this avoid duplicating the whole section in memory ('history')
for the sole purpose of computing the CRC.

this should be done also be done in dict_sect_pfc.rs, sequence.rs
and bitmap.rs, and the dependency on crc_any could then be dropped.
