### 2023-02-12 9b61eb3
raise error instead of panicking if Header is not valid UTF-8
