### 2023-02-12 819e0dd
more accurate error message
(it could be that the property 'length' was present,
but could not be parsed to a usize.
