### 2023-02-12 60559c4
use the Guard pattern for parsing Header's length
see https://en.wikipedia.org/wiki/Guard_%28computer_science%29
This is really a matter of taste, but I find it more readable.

Most of the lines "changed" in this commit are actually only reindented.
The only significant changes are at the beginning
(early return if length is invalid and missing)
and at the end (no more "else return error").
