+++ src/header.rs
@@ -28,47 +28,45 @@ impl Header {
             return Err(Error::new(InvalidData, "Headers currently only support the NTriples format"));
         }
 
-        let length = header_ci.get("length").and_then(|v| v.parse::<usize>().ok());
+        let length = header_ci
+            .get("length")
+            .and_then(|v| v.parse::<usize>().ok())
+            .ok_or_else(|| Error::new(InvalidData, "Header's length is missing or invalid"))?;
 
-        if let Some(length) = length {
-            let mut body_buffer: Vec<u8> = vec![0; length];
-            reader.read_exact(&mut body_buffer)?;
-            let mut body = BTreeSet::new();
+        let mut body_buffer: Vec<u8> = vec![0; length];
+        reader.read_exact(&mut body_buffer)?;
+        let mut body = BTreeSet::new();
 
-            for line_slice in body_buffer.split(|b| b == &b'\n') {
-                let line =
-                    str::from_utf8(line_slice).map_err(|_| Error::new(InvalidData, "Header is not UTF-8"))?;
-                if let Ok(Some(triple)) = triple_line(line) {
-                    let subject = match triple.subject {
-                        ntriple::Subject::IriRef(iri) => Id::Named(iri),
-                        ntriple::Subject::BNode(id) => Id::Blank(id),
-                    };
+        for line_slice in body_buffer.split(|b| b == &b'\n') {
+            let line = str::from_utf8(line_slice).map_err(|_| Error::new(InvalidData, "Header is not UTF-8"))?;
+            if let Ok(Some(triple)) = triple_line(line) {
+                let subject = match triple.subject {
+                    ntriple::Subject::IriRef(iri) => Id::Named(iri),
+                    ntriple::Subject::BNode(id) => Id::Blank(id),
+                };
 
-                    let ntriple::Predicate::IriRef(predicate) = triple.predicate;
+                let ntriple::Predicate::IriRef(predicate) = triple.predicate;
 
-                    let object = match triple.object {
-                        ntriple::Object::IriRef(iri) => Term::Id(Id::Named(iri)),
-                        ntriple::Object::BNode(id) => Term::Id(Id::Blank(id)),
-                        ntriple::Object::Lit(lit) => Term::Literal(match lit.data_type {
-                            ntriple::TypeLang::Lang(lan) => Literal::new_lang(lit.data, lan),
-                            ntriple::TypeLang::Type(data_type) => {
-                                if data_type == "http://www.w3.org/2001/XMLSchema#string" {
-                                    Literal::new(lit.data)
-                                } else {
-                                    Literal::new_typed(lit.data, data_type)
-                                }
+                let object = match triple.object {
+                    ntriple::Object::IriRef(iri) => Term::Id(Id::Named(iri)),
+                    ntriple::Object::BNode(id) => Term::Id(Id::Blank(id)),
+                    ntriple::Object::Lit(lit) => Term::Literal(match lit.data_type {
+                        ntriple::TypeLang::Lang(lan) => Literal::new_lang(lit.data, lan),
+                        ntriple::TypeLang::Type(data_type) => {
+                            if data_type == "http://www.w3.org/2001/XMLSchema#string" {
+                                Literal::new(lit.data)
+                            } else {
+                                Literal::new_typed(lit.data, data_type)
                             }
-                        }),
-                    };
+                        }
+                    }),
+                };
 
-                    body.insert(Triple::new(subject, predicate, object));
-                }
+                body.insert(Triple::new(subject, predicate, object));
             }
-
-            Ok(Header { format: header_ci.format, length, body })
-        } else {
-            Err(Error::new(InvalidData, "Header's length is missing or invalid"))
         }
+
+        Ok(Header { format: header_ci.format, length, body })
     }
 }
 
