### 2023-02-12 2dea620
make packed_data shareable across threads
this prevents the costly copy of data in the CRC thread
