### 2023-02-12 0b2ace2
ensure that FourDictSect.validate() is used before using
the trick consists in having FourSectDict::read return a different type,
UnvalidatedFourSectDict, which can then be converted to FourSectDict
through calling its validate() method.

In addition to being safer (you can't use the FourSectDict until it has
been validated), this makes DictSect and FourSectDict more memory
efficient, because they don't need to store the JoinHandle anymore.
