### 2023-02-13 43b9c85
Simpler return value
