### 2023-02-13 8f017ec
no need to duplicate history
full_words is not used after that, so we can reuse it directly instead
