### 2023-02-13 197f78d
use crate crc instead of crc_any in sequence.rs
following the change made in commit 78ab0c9
