### 2023-02-13 6e23da6
no need to duplicate history in bitmap.rs
full_words is not used after that, so we can reuse it directly
