### 2023-02-13 cfc0062
use crate crc instead of crc_any in bitmap.rs
following the change made in commit 78ab0c9
the dependency on crc-any can now be dropped
