+++ Cargo.toml
@@ -15,7 +15,6 @@ all-features = true
 [dependencies]
 bytesize = "1.1.0"
 crc = "3.0"
-crc-any = "2.3"
 iref = "2.2"
 langtag = "^0.3.2"
 ntriple = "^0.1.1"
