### 2023-02-13 159ee51
use crate crc instead of crc_any in bitmap.rs
following the change made in commit 78ab0c9
the dependency on crc-any can now be dropped
