### 2023-02-13 e31b0a9
Improve documentation and code examples. Re-export sophia as hdt::sophia. Release 0.0.13-alpha.1.
