+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "hdt"
-version = "0.0.13-alpha.0"
+version = "0.0.13-alpha.1"
 repository = "https://github.com/konradhoeffner/hdt"
 authors = ["Tim Baccaert <tbaccaer@vub.be>", "Konrad Höffner"]
 license = "MIT"
@@ -8,6 +8,9 @@ description = "Library for the Header Dictionary Triples (HDT) RDF compression f
 keywords = ["rdf", "hdt", "compression", "file-format"]
 categories = ["compression", "filesystem", "parsing", "web-programming"]
 edition = "2021"
+
+[package.metadata."docs.rs"]
+all-features = true
 
 [dependencies]
 bytesize = "1.1.0"
