### 2023-02-15 1671aeb
borrow_term() is the expected way to "lend" a term
(i.e. when you want to keep ownership of it,
but need to pass it to a function expecting T: Term)
https://docs.rs/sophia_api/0.8.0-alpha.0/sophia_api/term/trait.Term.html#tymethod.borrow_term

NB:
1. this must be explained and documented in the Sophia book
   https://pchampin.github.io/sophia_rs/
2. in the near future, i am considering renaming this method,
   maybe to "proxy()"
