### 2023-02-16 b3d1b1a
Apply a clippy suggestion.
