+++ src/lib.rs
@@ -50,6 +50,9 @@
 #![warn(missing_docs)]
 #![warn(clippy::pedantic)]
 #![warn(clippy::cargo)]
+#![warn(clippy::str_to_string)]
+#![warn(clippy::print_stdout)]
+#![warn(clippy::print_stderr)]
 #![allow(clippy::unnecessary_cast)]
 #![allow(clippy::must_use_candidate)]
 #![allow(clippy::missing_errors_doc)]
