### 2023-02-17 cea23b3
Check formatting and lint in CI and add CI badge.
