+++ .github/workflows/test.yml
@@ -1,29 +1,0 @@
-name: Test
-
-on:
-  - push
-  - pull_request
-
-jobs:
-  build:
-    strategy:
-      matrix:
-        os: [ubuntu-latest]
-        rust-version: [nightly]
-        include:
-          - os: macos-latest
-            rust-version: nightly
-          - os: windows-latest
-            rust-version: nightly
-    runs-on: ${{ matrix.os }}
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Rust ${{ matrix.rust-version }}
-        uses: actions-rs/toolchain@v1
-        with:
-          toolchain: ${{ matrix.rust-version }}
-          override: true
-      - name: Build
-        run: cargo build --verbose
-      - name: Run tests
-        run: cargo test --verbose
