+++ .github/workflows/lint_and_test.yml
@@ -1,0 +1,29 @@
+name: Lint and Test
+
+on:
+  - push
+  - pull_request
+
+jobs:
+  build:
+    strategy:
+      matrix:
+        include:
+          - os: ubuntu-latest
+          #- os: macos-latest
+          #- os: windows-latest
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: actions/checkout@v3
+      - name: Install Rust
+        uses: dtolnay/rust-toolchain@nightly
+        with:
+          components: rustfmt,clippy
+      - name: fmt 
+        run: cargo fmt -- --check
+      - name: clippy
+        run: cargo clippy --no-deps --all-features
+      - name: build
+        run: cargo build
+      - name: test
+        run: cargo test --all-features
