### 2023-02-19 a430efa
hdt::SimpleTriple now uses Arc<str> instead of MownStr
(also, cargo fmt slightly changed the layout)

This is a first step toward improving performance,
although at the moment it makes things slightly slower.

The idea in the future is to be able to reuse the same Arc<str>,
rather than allocating a new one each time.
