### 2023-02-19 85dd26c
HdtGraph now use HdtTerm, an Arc<str>-based Term implementation
