### 2023-02-19 6fdc9ee
HdtGraph now use HdtTerm, an Arc<str>-based Term implementation
