+++ src/hdt_graph.rs
@@ -5,15 +5,16 @@ use crate::hdt::Hdt;
 use crate::triples::{Id, ObjectIter, PredicateIter, PredicateObjectIter, SubjectIter, TripleId};
 use log::debug;
 
-use mownstr::MownStr;
 use sophia::api::graph::{GTripleSource, Graph};
 
-use sophia::api::term::FromTerm;
-use sophia::api::term::{matcher::TermMatcher, BnodeId, IriRef, LanguageTag, SimpleTerm, Term};
+use sophia::api::term::{matcher::TermMatcher, BnodeId, IriRef, LanguageTag, Term};
 use std::convert::Infallible;
 use std::io::{self, Error, ErrorKind};
 use std::iter;
 use std::sync::Arc;
+
+mod term;
+pub use term::HdtTerm;
 
 /// Adapter to use HDT as a Sophia graph.
 pub struct HdtGraph {
@@ -30,18 +31,36 @@ impl HdtGraph {
         self.hdt.size_in_bytes()
     }
 
-    fn id_term(&self, id: Id, kind: &'static IdKind) -> SimpleTerm<'static> {
-        let s = MownStr::from(self.hdt.dict.id_to_string(id, kind).unwrap());
-        auto_term(s).unwrap()
+    fn id_term(&self, id: Id, kind: &'static IdKind) -> HdtTerm {
+        auto_term(self.hdt.dict.id_to_string(id, kind).unwrap().into()).unwrap()
         // TODO: optimize by excluding cases depending on the id kind
         //IriRef::new_unchecked(MownStr::from(s)).into_term()
+    }
+
+    /// Return
+    /// - Ok(None) if tm is not a constant matcher
+    /// - Ok(Some(term, term_id))) if tm is a constant matcher
+    /// - Err(()) if tm is a constant matcher, but its only term is not in this graph
+    fn unpack_matcher<T: TermMatcher>(&self, tm: T, kind: &IdKind) -> Result<Option<(HdtTerm, Id)>, ()> {
+        match tm.constant() {
+            Some(t) => match HdtTerm::try_from(t.borrow_term()) {
+                Some(t) => {
+                    let id = self.hdt.dict.string_to_id(&term_string(&t), kind);
+                    if id == 0 {
+                        return Err(());
+                    }
+                    Ok(Some((t, id)))
+                }
+                None => Err(()),
+            },
+            None => Ok(None),
+        }
     }
 }
 
 /// Create the correct Sophia term for a given resource string.
 /// Slow, use the appropriate method if you know which type (Literal, URI, or blank node) the string has.
-fn auto_term(s: Arc<str>) -> io::Result<SimpleTerm<'static>> {
-    let s = MownStr::from(s.to_string());
+fn auto_term(s: Arc<str>) -> io::Result<HdtTerm> {
     match s.chars().next() {
         None => Err(Error::new(ErrorKind::InvalidData, "empty input")),
         Some('"') => match s.rfind('"') {
@@ -50,17 +69,16 @@ fn auto_term(s: Arc<str>) -> io::Result<SimpleTerm<'static>> {
                 format!("missing right quotation mark in literal string {s}"),
             )),
             Some(index) => {
-                let lex = &s[1..index];
+                let lex = Arc::from(&s[1..index]);
                 let rest = &s[index + 1..];
                 // literal with no language tag and no datatype
                 if rest.is_empty() {
-                    return Ok(lex.into_term());
+                    return Ok(HdtTerm::LiteralDatatype(lex, term::XSD_STRING.clone()));
                 }
-                let lex = MownStr::from_str(lex);
                 // either language tag or datatype
                 if let Some(tag_index) = rest.find('@') {
[…]
