### 2023-02-19 adbaf6e
simplify tests
- use [t] instead of Some(t) as term matchers (more concise)
- use t.borrow_term() instead of t.clone() (lest costly)
- use &str as term (no need to convert it to SimpleTerm)
