### 2023-02-20 6068f43
Remove unused objects_with_sp method.
