### 2023-02-22 bfd0ed4
Benchmark Sophia graph as well. Make wrapped HDT instance of Sophia graph public.
