+++ benches/bench.rs
@@ -1,7 +1,12 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 use hdt::triples::*;
 use hdt::Hdt;
+use hdt::HdtGraph;
 use hdt::IdKind;
+use sophia::api::graph::Graph;
+use sophia::api::term::matcher::Any;
+use sophia::api::term::IriRef;
+use sophia::api::term::SimpleTerm;
 use std::fs::File;
 
 fn bench_query(c: &mut Criterion) {
@@ -12,40 +17,61 @@ fn bench_query(c: &mut Criterion) {
     //let file = File::open("tests/resources/lscomplete2015.hdt").expect("error opening file");
     //let file = File::open("tests/resources/snikmeta.hdt").expect("error opening file");
     let hdt = Hdt::new(std::io::BufReader::new(file)).unwrap();
-    let triples = &hdt.triples;
     let vincent_id = hdt.dict.string_to_id(vincent, &IdKind::Subject);
     let type_id = hdt.dict.string_to_id(type_, &IdKind::Predicate);
     let person_id = hdt.dict.string_to_id(person, &IdKind::Object);
-    let twp = |s, p, o| hdt.triples_with_pattern(s, p, o);
+    let graph = HdtGraph::new(hdt);
+    let vincent_term = SimpleTerm::Iri(IriRef::new_unchecked(vincent.into()));
+    let type_term = SimpleTerm::Iri(IriRef::new_unchecked(type_.into()));
+    let person_term = SimpleTerm::Iri(IriRef::new_unchecked(person.into()));
+    let twp = |s, p, o| graph.hdt.triples_with_pattern(s, p, o);
+    let triples = &graph.hdt.triples;
 
     // count to prevent optimizing away function call
-    //let mut group = c.benchmark_group("S??");
-    let mut group = c.benchmark_group("query");
-    group.bench_function("1 triple IDs (vincent, ?, ?)", |b| {
+    let mut group = c.benchmark_group("S??");
+    //let mut group = c.benchmark_group("query");
+    group.bench_function("1.1 (vincent, ?, ?) triple IDs", |b| {
         b.iter(|| SubjectIter::with_pattern(triples, &TripleId::new(vincent_id, 0, 0)).count())
     });
-    group.bench_function("2 str triples (vincent, ?, ?)", |b| b.iter(|| twp(Some(vincent), None, None).count()));
-    //group.finish();
-    //let mut group = c.benchmark_group("?P?");
+    group.bench_function("1.2 (vincent, ?, ?) str triples", |b| b.iter(|| twp(Some(vincent), None, None).count()));
+    group.bench_function("1.3 (vincent, ?, ?) Sophia triples", |b| {
+        b.iter(|| graph.triples_matching(Some(&vincent_term), Any, Any).count())
+    });
+    group.finish();
+
+    let mut group = c.benchmark_group(format!("?P? {} triples", PredicateIter::new(triples, type_id).count()));
     group.sample_size(10);
-    group.bench_function("3 triple IDs (?, type, ?)", |b| b.iter(|| PredicateIter::new(triples, type_id).count()));
-    group.bench_function("4 str triples (?, type, ?)", |b| b.iter(|| twp(None, Some(type_), None).count()));
-    //group.finish();
-    //let mut group = c.benchmark_group("??O");
-    group
-        .bench_function("5 triple IDs (?, ?, person)", |b| b.iter(|| ObjectIter::new(triples, person_id).count()));
-    group.bench_function("6 str triples (?, ?, person)", |b| b.iter(|| twp(None, None, Some(person)).count()));
-    //group.finish();
-    //let mut group = c.benchmark_group("?PO");
+    group.bench_function("2.1 (?, type, ?) triple IDs", |b| {
[…]
