### 2023-02-22 0072b8b
Simplify object ierator.
