### 2023-02-22 5ec68fb
Bench triples().
