### 2023-02-23 7da5bd0
Temporarily use a fork of Criterion with a quick mode bugfix.
