+++ Cargo.toml
@@ -43,4 +43,6 @@ opt-level = 1
 [dev-dependencies]
 pretty_assertions = "1.3"
 env_logger = { version = "0.10", default-features = false, features = ["auto-color"] }
-criterion = { version = "0.4.0", default-features = false, features = ["cargo_bench_support", "html_reports"] }
+#criterion = { version = "0.4.0", default-features = false, features = ["cargo_bench_support", "html_reports"] }
+# bug fix for queries longer than 5 seconds in quick mode
+criterion = { version = "0.4.0", default-features = false, features = ["cargo_bench_support", "html_reports"], git = "https://github.com/konradhoeffner/criterion.rs" }
