### 2023-02-23 24e21f0
Experiment with a string pool instead of arc or mownstr.
