+++ src/hdt.rs
@@ -5,7 +5,6 @@ use crate::triples::{ObjectIter, PredicateIter, PredicateObjectIter, SubjectIter
 use crate::FourSectDict;
 use bytesize::ByteSize;
 use log::{debug, error};
-use mownstr::MownStr;
 use std::io;
 use std::iter;
 use thiserror::Error;
@@ -22,7 +21,7 @@ pub struct Hdt {
     pub triples: TriplesBitmap,
 }
 
-type StringTriple<'a> = (MownStr<'a>, MownStr<'a>, MownStr<'a>);
+type StringTriple<'a> = (&'a str,&'a str, &'a str);
 
 /// The error type for the `translate_id` method.
 #[derive(Error, Debug)]
@@ -60,13 +59,13 @@ impl Hdt {
     pub fn size_in_bytes(&self) -> usize {
         self.dict.size_in_bytes() + self.triples.size_in_bytes()
     }
-
+/*
     /// Don't use this for many triples with shared values as you won't benefit from deduplication.
     fn translate_id(&self, t: TripleId) -> Result<StringTriple, TranslateErr> {
         let s = self.dict.id_to_string(t.subject_id, &IdKind::Subject).map_err(|e| TranslateErr { e, t })?;
         let p = self.dict.id_to_string(t.predicate_id, &IdKind::Predicate).map_err(|e| TranslateErr { e, t })?;
         let o = self.dict.id_to_string(t.object_id, &IdKind::Object).map_err(|e| TranslateErr { e, t })?;
-        Ok((s.into(), p.into(), o.into()))
+        Ok((&s, &p, &o))
     }
 
     /// An iterator visiting *all* triples as strings in order.
@@ -76,7 +75,7 @@ impl Hdt {
     pub fn triples(&self) -> impl Iterator<Item = StringTriple> + '_ {
         // TODO deduplicate
         self.triples.into_iter().map(|id| self.translate_id(id).unwrap())
-    }
+    }*/
 
     /// Get all subjects with the given property and object (?PO pattern).
     /// Use this over `triples_with_pattern(None,Some(p),Some(o))` if you don't need whole triples.
@@ -102,38 +101,42 @@ impl Hdt {
     /// Get all triples that fit the given triple patterns, where `None` stands for a variable.
     /// For example, `triples_with_pattern(None, Some(p), Some(o)` answers an ?PO pattern.
     pub fn triples_with_pattern<'a>(
-        &'a self, sp: Option<&'a str>, pp: Option<&'a str>, op: Option<&'a str>,
-    ) -> Box<dyn Iterator<Item = StringTriple<'a>> + '_> {
-        let xso = sp.map(|s| (MownStr::from_str(s), self.dict.string_to_id(s, &IdKind::Subject)));
-        let xpo = pp.map(|p| (MownStr::from_str(p), self.dict.string_to_id(p, &IdKind::Predicate)));
-        let xoo = op.map(|o| (MownStr::from_str(o), self.dict.string_to_id(o, &IdKind::Object)));
+        &'a self, sp: Option<&'a str>, pp: Option<&'a str>, op: Option<&'a str>, pool: &'a mut Vec<String>
+    ) -> Vec<StringTriple<'a>> {
+        let xso = sp.map(|s| (s, self.dict.string_to_id(s, &IdKind::Subject)));
+        let xpo = pp.map(|p| (p, self.dict.string_to_id(p, &IdKind::Predicate)));
+        let xoo = op.map(|o| (o, self.dict.string_to_id(o, &IdKind::Object)));
+        let mut result = Vec::<_>::new();
         if [&xso, &xpo, &xoo].into_iter().flatten().any(|x| x.1 == 0) {
             // at least one term does not exist in the graph
[…]
