+++ Cargo.toml
@@ -24,10 +24,11 @@ sophia = { version="0.8.0-alpha.0", optional = true }
 sucds = "0.6.0"
 thiserror = "1.0.37"
 log = "0.4"
-mownstr = "0.2.0"
+#mownstr = "0.2.0"
 
 [features]
-default = ["sophia"]
+#default = ["sophia"]
+default = []
 sophia = ["dep:sophia"]
 
 [[bench]]
