### 2023-02-24 9b57984
Use MownStr borrow instead of clone with a local patched version.
