### 2023-03-02 #5114 @remram44
I am reviewing the code and will attempt to go through the benchmark as well.

-----

My concern at this point is with the [Substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) requirement. While it is clear that some effort has gone into this and that it is useful to the community, it seems on the line with regards to JOSS's requirements.

This project was forked off of tbaccaer's original on September 27, the first version was tagged and pushed to crates.io on November 14, and it was submitted to JOSS on January 18. The original (pre-fork) was implemented in a single week.

This is very close to the hard minimum for JOSS of three months of work.

Because it reads a single file format invented and published by somebody else, and already has an integration with the [Sophia toolkit for Linked Data in Rust](https://github.com/pchampin/sophia_rs), it looks like an incremental improvement to their project more than stand-alone research software. The only user listed in the paper is an application from the author, where this library is used together with Sophia. It seems unlikely to me that this library would gain significant notoriety to the point of being ”likely to be cited”.

I also worry about whether it is "feature-complete" as it is currently undergoing significant API changes for version 0.0.13 currently in alpha.

If my co-reviewer is ok with this I will not oppose publication but I wanted to raise this concern.
