### 2023-03-02 #5114 @lazear
It seems to me that tbaccaer is an author on the paper, and that there has been enough effort to qualify for publication - at least on the software portion. While I know JOSS papers should be brief, I think the paper in its current form is shorter than most of the recent JOSS papers I have looked at. I suggested addition of some examples in the linked issue.

I can't claim to be an expert (or even knowledgeable) about RDF/semantic web technologies, so I don't want to preclude publication based on likelihood of being cited - but I do share some of @remram44's concerns about whether this is currently in a state where it could be used by other applications besides Sophia. It seems that the library can currently be used to read the HDT format, but not write back to disk - this seems like something that would be a very useful feature.
