### 2023-03-03 #5114 @osorensen
Thanks to both @remram44 and @lazear for your comments.

I suggest that @KonradHoeffner starts by addressing the following issue from @remram44

> I also worry about whether it is "feature-complete" as it is currently undergoing significant API changes for version 0.0.13 currently in alpha.

Furthermore, please also expand the paper according to @lazear's suggestions.

Once done, I suggest @KonradHoeffner reports back, and then let's the reviewers have a look at the updated versions of the software as well as the updated paper.
