### 2023-03-03 #5114 @KonradHoeffner
I can and will absolutely expand on the paper if you want! I just went after the submission guidelines of 250-1000 words so I thought with 485 I'm comfortably in the middle but if you think it would profit from more examples, it is no problem for me to add more examples and motivate the significance of it better.

As for the other points:

* Tim Baccaert (@timplication) and me are indeed both contributors and he is listed on the paper with "these authors contributed equally".
* The original, incomplete version (base of the fork) at <https://github.com/timplication/hdt-rs/> is not available anymore and this is the quasi successor of it. The paper is thus not only about the improvements of the fork (done by mostly me) but also about the original version (done by @timplication alone).
* In my admittedly biased opinion this paper has a very high likelihood of being cited, as it is currently the only available Rust library to query HDT.
* While HDT might seem like just another RDF format/serialization, it being a binary compressed format enables large memory savings.
* Writing back HDT to disk is of limited use because the format is read-only due to the compression. While the Java and C++ versions support that, it is mostly used to convert other RDF serializations to HDT with command line applications like rdf2hdt of hdt-cpp in addition to the library. But if you want to convert for example RDF Turtle to HDT you can just use hdt-cpp for that, so I think this is not needed that much, while you can't use a C++ or Java library in Rust that easily.
* The alpha version is due to the optional Sophia dependency currently being in alpha which requires extensive refactoring of the Sophia adapter. However as soon as the Sophia non-alpha version is released, this library will follow as well. The idea was to get a head start on doing the refactoring so it does not need to be done later.
* There are indeed API changes in the alpha version and some performance optimizations but they do not add any new features, they just make querying more comfortable by supplying a single function for all triple patterns where before there were multiple functions. Both the stable and the alpha version has all the features and passes all the tests, linting and so on.
