+++ paper.md
@@ -63,6 +63,6 @@ Extensions such as HDT++ [@serializingrdf] or iHDT++ [@ihdt] are currently not s
 # Acknowledgements
 
 We thank Pierre-Antoine Champin for explaining the details of Sophia and for creating [the benchmark suite](https://github.com/pchampin/sophia_benchmark) 
-that the [HDT benchmarks](https://github.com/KonradHoeffner/hdt_benchmark) are based on.
+that the [HDT benchmarks](https://github.com/KonradHoeffner/hdt_benchmark) are based on and for the thorough code review.
 
 # References
