### 2023-03-08 2617927
More preliminaries and examples.
