### 2023-03-08 726e9be
Yet more preliminaries and examples.
