+++ paper.bib
@@ -140,3 +140,14 @@ url = {https://www2020devtrack.github.io/site/schedule}
   year={2020},
   publisher={IOS Press}
 }
+
+@article{dbpedia,
+  title={{D}bpedia -- {A} large-scale, multilingual knowledge base extracted from wikipedia},
+  author={Lehmann, Jens and Isele, Robert and Jakob, Max and Jentzsch, Anja and Kontokostas, Dimitris and Mendes, Pablo N and Hellmann, Sebastian and Morsey, Mohamed and Van Kleef, Patrick and Auer, S{\"o}ren and others},
+  journal={Semantic web},
+  volume={6},
+  number={2},
+  pages={167--195},
+  year={2015},
+  publisher={IOS Press}
+}
