+++ paper.md
@@ -30,6 +30,102 @@ We present the Rust library hdt-rs (named "hdt" in the context of Rust libraries
 This allows writing high-performance Rust applications that load and query HDT files using triple patterns.
 Existing Rust applications using the Sophia [@sophia] library can easily and drastically reduce their RAM usage by using the provided Sophia HDT adapter.
 
+# Preliminaries
+
+## RDF
+
+The *Resource Description Framework* (RDF) is a data model that represents information using *triples*, each consisting of a *subject*, *predicate* and *object*.
+A set of triples is called an *RDF graph*, where the subjects and objects can be visualized as nodes and the predicates as labeled, directed edges.
+Predicates are always IRIs, however subjects and objects can also be *blank nodes* and objects can also be *literals*.
+There are multiple text-based RDF serialization formats with different compromises between verbosity, ease of automatic processing and human readability.
+For example, the N-Triples serialization of the fact "the mayor of Leipzig is Burkhard Jung" from DBpedia [@dbpedia] is:
+
+```ntriples
+<http://dbpedia.org/resource/Leipzig> <http://dbpedia.org/ontology/mayor>
+    (no linebreak) <http://dbpedia.org/resource/Burkhard_Jung> .
+```
+
+## Triple Patterns
+
+*Triple patterns* allow matching a subset of a graph.
+Each part of the pattern is either a constant or a variable, resulting in eight different types of triple patterns. 
+We denote the triple pattern type with all constants as SPO (matching one or zero triples) and the type with all constants with ??? (matching all triples in the graph).
+The other triple patterns are denoted analogously.
+
+## Header Data Triples
+While text-based RDF serialization formats can be read by humans, they are too verbose to be practical on large graphs.
+The serialized size of a graph can be drastically lowered by using the Header Data Triples (HDT) binary RDF format, which can be loaded into memory in compressed form while still allowing efficient queries.
+
+![The Bitmap Triples structure represents the adjacency matrix of the RDF graph as a tree.
+Image source and further information in @hdt2012.
+\label{fig:bt}](img/bt.png){ width=100% }
+
+All patterns with constant subject (SPO, SP?, SO? and S??) as well as ??? are answered using the Bitmap Triples structure, see \autoref{fig:bt}.
+
+![The HDT *Focused on Querying* (HDT-FoQ) extension allows efficient queries with ?PO, ?P? and ??O patterns.
+Image source and further information in @hdt2012.
+\label{fig:foq}](img/hdt-foq.png){ width=50% }
+
+# Examples
+
+## Add the dependency to a Rust application
+
+```bash
+$ cargo add hdt
+```
+
+## Load an HDT file
+
+```rust
+use hdt::Hdt;
+use std::{fs::File,io::BufReader};
+
+let f = File::open("example.hdt").expect("error opening file");
+let hdt = Hdt::new(BufReader::new(f)).expect("error loading HDT");
+```
+## Query SP? pattern
+
+Find the mayor of Leipzig from DBpedia using an SP? triple pattern:
+
+```rust
+let mayors = hdt.triples_with_pattern(
+    Some("http://dbpedia.org/resource/Leipzig"),
+    Some("http://dbpedia.org/ontology/mayor"),
+    None);
+println!("{:?}", mayors.collect::<Vec<_>>());
+```
+
+## Query ?PO pattern
+
+Which city has Burkhard Jung as the mayor?
+
+```rust
+let mayors = hdt.triples_with_pattern(
+    None,
+    Some("http://dbpedia.org/ontology/mayor"),
+    Some("http://dbpedia.org/resource/Leipzig"));
[…]
