+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'HDT: A Rust library for the Header Data Triples binary RDF compression format'
+title: 'HDT: A Rust library for the Header Dictionary Triples binary RDF compression format'
 tags:
   - Rust
   - HDT
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Summary
 
-We present the Rust library hdt-rs (named "hdt" in the context of Rust libraries, such as [on crates.io](https://crates.io/crates/hdt)) for the Header Data Triples (HDT) binary RDF compression format described by @hdt2012 and @hdt2013.
+We present the Rust library hdt-rs (named "hdt" in the context of Rust libraries, such as [on crates.io](https://crates.io/crates/hdt)) for the Header Dictionary Triples (HDT) binary RDF compression format.
 This allows writing high-performance Rust applications that load and query HDT files using triple patterns.
 Existing Rust applications using the Sophia [@sophia] library can easily and drastically reduce their RAM usage by using the provided Sophia HDT adapter.
 
@@ -52,15 +52,18 @@ Each part of the pattern is either a constant or a variable, resulting in eight 
 We denote the triple pattern type with all constants as SPO (matching one or zero triples) and the type with all constants with ??? (matching all triples in the graph).
 The other triple patterns are denoted analogously.
 
-## Header Data Triples
+## Header Dictionary Triples
 While text-based RDF serialization formats can be read by humans, they are too verbose to be practical on large graphs.
-The serialized size of a graph can be drastically lowered by using the Header Data Triples (HDT) binary RDF format, which can be loaded into memory in compressed form while still allowing efficient queries.
+The serialized size of a graph can be drastically lowered by using the Header Dictionary Triples (HDT) binary RDF format, which can be loaded into memory in compressed form while still allowing efficient queries.
+
 
 ![The Bitmap Triples structure represents the adjacency matrix of the RDF graph as a tree.
 Image source and further information in @hdt2012.
 \label{fig:bt}](img/bt.png){ width=100% }
 
-All patterns with constant subject (SPO, SP?, SO? and S??) as well as ??? are answered using the Bitmap Triples structure, see \autoref{fig:bt}.
+All patterns with constant subject (SPO, SP?, SO? and S??) as well as ??? are answered using the Bitmap Triples structure, see \autoref{fig:bt}, while the other
+patterns are answered using HDT-FoQ, see \autoref{fig:foq}.
+As HDT is a very complex format, we recommend referring to @hdt2012 and @hdt2013 for a comprehensive documentation.
 
 ![The HDT *Focused on Querying* (HDT-FoQ) extension allows efficient queries with ?PO, ?P? and ??O patterns.
 Image source and further information in @hdt2012.
@@ -138,8 +141,9 @@ Longterm availability of SPARQL endpoints is often compromised [@readyforaction]
 To insulate against such problems, Semantic Web applications may integrate and query an RDF graph using libraries such as Apache Jena [@jena] for Java,
 RDFlib [@rdflib] for Python, librdf [@librdf] for C or Sophia [@sophia] for Rust.
 However those libraries do not scale to large RDF graphs due to their excessive memory usage, see \autoref{fig:benchmark},
-which can be drastically lowered by using the Header Data Triples (HDT) binary RDF format, which can be loaded into memory in compressed form while still allowing efficient queries.
-Implementations existed for C++ [@hdtcpp] and Java [@hdtjava] but not for Rust, a popular modern, statically typed high-level programming language that allows writing performant software while still ensuring memory safety,
[…]
