### 2023-03-08 #5114 @lazear
I think this looks great, and does a better job of motivating hdt-rs and explaining it's use - only nit is "Predicates are always IRIs" - IRI isn't previously defined.
