+++ paper.bib
@@ -151,3 +151,10 @@ url = {https://www2020devtrack.github.io/site/schedule}
   year={2015},
   publisher={IOS Press}
 }
+
+@book{frontcoding,
+  title={Managing gigabytes: Compressing and indexing documents and images},
+  author={Witten, Ian H and Witten, Ian H and Moffat, Alistair and Bell, Timothy C and Bell, Timothy C and Fox, Ed and Bell, Timothy C},
+  year={1999},
+  publisher={Morgan Kaufmann}
+}
