+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "hdt"
-version = "0.0.13-alpha.2"
+version = "0.0.13-alpha.3"
 repository = "https://github.com/konradhoeffner/hdt"
 authors = ["Tim Baccaert <tbaccaer@vub.be>", "Konrad Höffner"]
 license = "MIT"
@@ -43,6 +43,4 @@ opt-level = 1
 [dev-dependencies]
 pretty_assertions = "1.3"
 env_logger = { version = "0.10", default-features = false, features = ["auto-color"] }
-#criterion = { version = "0.4.0", default-features = false, features = ["cargo_bench_support", "html_reports"] }
-# bug fix for queries longer than 5 seconds in quick mode
-criterion = { version = "0.4.0", default-features = false, features = ["cargo_bench_support", "html_reports"], git = "https://github.com/konradhoeffner/criterion.rs" }
+criterion = { version = "0.4.0", default-features = false, features = ["cargo_bench_support", "html_reports"] }
