+++ Cargo.toml
@@ -19,8 +19,8 @@ iref = "2.2"
 langtag = "^0.3.2"
 ntriple = "^0.1.1"
 #rayon = "1.5.3"
-rsdict = "0.0.6"
-sophia = { version="0.8.0-alpha.0", optional = true }
+rsdict = { version = "0.0.6", features = ["simd"] }
+sophia = { version = "0.8.0-alpha.0", optional = true }
 sucds = "0.6.0"
 thiserror = "1.0.37"
 log = "0.4"
