+++ paper.bib
@@ -110,7 +110,7 @@ url = {https://www2020devtrack.github.io/site/schedule}
 
 @InProceedings{readyforaction,
   author="Buil-Aranda, Carlos and Hogan, Aidan and Umbrich, J{\"u}rgen and Vandenbussche, Pierre-Yves",
-  editor="Alani, Harith and Kagal, Lalana and Fokoue, Achille and Groth, Paul and Biemann, Chris and Parreira, Josiane Xavier and Aroyo, Lora and Noy, Natasha and Welty, Chris and Janowicz, Krzysztof",
+  noeditor="Alani, Harith and Kagal, Lalana and Fokoue, Achille and Groth, Paul and Biemann, Chris and Parreira, Josiane Xavier and Aroyo, Lora and Noy, Natasha and Welty, Chris and Janowicz, Krzysztof",
   title="{SPARQL} Web-Querying Infrastructure: Ready for Action?",
   booktitle="The Semantic Web -- ISWC 2013",
   year="2013",
@@ -142,7 +142,7 @@ url = {https://www2020devtrack.github.io/site/schedule}
 }
 
 @article{dbpedia,
-  title={{D}bpedia -- {A} large-scale, multilingual knowledge base extracted from wikipedia},
+  title={{DB}pedia -- {A} large-scale, multilingual knowledge base extracted from wikipedia},
   author={Lehmann, Jens and Isele, Robert and Jakob, Max and Jentzsch, Anja and Kontokostas, Dimitris and Mendes, Pablo N and Hellmann, Sebastian and Morsey, Mohamed and Van Kleef, Patrick and Auer, S{\"o}ren and others},
   journal={Semantic web},
   volume={6},
@@ -157,4 +157,13 @@ url = {https://www2020devtrack.github.io/site/schedule}
   author={Witten, Ian H and Witten, Ian H and Moffat, Alistair and Bell, Timothy C and Bell, Timothy C and Fox, Ed and Bell, Timothy C},
   year={1999},
   publisher={Morgan Kaufmann}
+}
+
+@inproceedings{waveletmatrix,
+  title={The Wavelet Matrix},
+  author={Claude, Francisco and Navarro, Gonzalo},
+  booktitle={String Processing and Information Retrieval: 19th International Symposium, SPIRE 2012, Cartagena de Indias, Colombia, October 21-25, 2012. Proceedings 19},
+  pages={167--179},
+  year={2012},
+  organization={Springer}
 }
