+++ paper.md
@@ -39,7 +39,7 @@ A set of triples is called an *RDF graph*, where the subjects and objects can be
 A predicate is always an *IRI* (Internationalized Resource Identifier), which is a generalization of an URI that permits additional characters.
 Subjects and objects can also be *blank nodes* and objects can also be *literals*.
 There are multiple text-based RDF serialization formats with different compromises between verbosity, ease of automatic processing and human readability.
-For example, the N-Triples serialization of the fact "the mayor of Leipzig is Burkhard Jung" from DBpedia [@dbpedia] is:
+For example, the N-Triples representation of the fact "the mayor of Leipzig is Burkhard Jung" from DBpedia [@dbpedia] is:
 
 ```ntriples
 <http://dbpedia.org/resource/Leipzig> <http://dbpedia.org/ontology/mayor>
@@ -49,8 +49,8 @@ For example, the N-Triples serialization of the fact "the mayor of Leipzig is Bu
 ## Triple Patterns
 
 *Triple patterns* allow matching a subset of a graph.
-Each part of the pattern is either a constant or a variable, resulting in eight different types of triple patterns. 
-We denote the triple pattern type with all constants as SPO (matching one or zero triples) and the type with all constants with ??? (matching all triples in the graph).
+Each part of the pattern is either a constant or a variable, resulting in eight different types. 
+We denote the pattern type with all constants as SPO (matching one or zero triples) and the type with all constants with ??? (matching all triples in the graph).
 The other triple patterns are denoted analogously.
 
 ## Header Dictionary Triples
@@ -73,6 +73,32 @@ As HDT is a very complex format, we recommend referring to @hdt2012 and @hdt2013
 Image source and further information in @hdt2012.
 \label{fig:foq}](img/hdt-foq.png){ width=50% }
 
+# Statement of need
+
+Semantic Web technologies have seen adoption by major tech companies in recent years
+but widespread use is still inhibited by a lack of freely available performant, accessible, robust and adaptable tooling [@semanticwebreview].
+SPARQL endpoints provide a standard publication channel and API to any RDF graph but they are not suitable for all use cases.
+On small graphs, there is a large relative overhead in both memory and CPU resources.
+On large graphs on the other hand, query complexity and shared access may cause an overload of the server, causing delayed or missed responses.
+Longterm availability of SPARQL endpoints is often compromised [@readyforaction], which impacts all applications depending on them.
+
+To insulate against such problems, Semantic Web applications may integrate and query an RDF graph using libraries such as Apache Jena [@jena] for Java,
+RDFlib [@rdflib] for Python, librdf [@librdf] for C or Sophia [@sophia] for Rust.
+However those libraries do not scale to large RDF graphs due to their excessive memory usage, see \autoref{fig:benchmark}.
+To complement hdt-cpp [@hdtcpp] and hdt-java [@hdtjava], we implement HDT in Rust, which is a popular modern, statically typed high-level programming language that allows writing performant software while still ensuring memory safety,
+which aligns with the challenges to the adoption of the Semantic Web.
+hdt-rs is used through the included Sophia adapter by the RickView [@rickview] RDF browser to publish large graphs, for example LinkedSpending [@linkedspending] at <https://linkedspending.aksw.org>,
+which previously suffered from frequent downtimes when based on a SPARQL endpoint.
+
+# Benchmark
+
[…]
