+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "hdt"
-version = "0.0.13-alpha.3"
+version = "0.0.13"
 repository = "https://github.com/konradhoeffner/hdt"
 authors = ["Tim Baccaert <tbaccaer@vub.be>", "Konrad Höffner"]
 license = "MIT"
@@ -18,7 +18,6 @@ crc = "3.0"
 iref = "2.2"
 langtag = "^0.3.2"
 ntriple = "^0.1.1"
-#rayon = "1.5.3"
 rsdict = { version = "0.0.6", features = ["simd"] }
 sophia = { version = "0.8.0-alpha.0", optional = true }
 sucds = "0.6.0"
