### 2023-03-10 872609c
Add cargo example. Part of #29.
