+++ paper.bib
@@ -167,3 +167,18 @@ url = {https://www2020devtrack.github.io/site/schedule}
   year={2012},
   organization={Springer}
 }
+
+@article{compressedindexes,
+author = {R. Perego and G. Pibiri and R. Venturini},
+journal = {IEEE Transactions on Knowledge and Data Engineering},
+title = {Compressed Indexes for Fast Search of Semantic Data},
+year = {2021},
+volume = {33},
+number = {09},
+issn = {1558-2191},
+pages = {3187-3198},
+doi = {10.1109/TKDE.2020.2966609},
+publisher = {IEEE Computer Society},
+address = {Los Alamitos, CA, USA},
+month = {sep}
+}
