### 2023-03-21 #5114 @KonradHoeffner
> I also worry about whether it is "feature-complete" as it is currently undergoing significant API changes for version 0.0.13 currently in alpha.

0.0.13 is now (since March 10) fully released (i.e. not alpha anymore), see <https://github.com/KonradHoeffner/hdt/releases/tag/0.0.13>.
