### 2023-03-21 1b2dc76
Add benchmark table and explanation.
