### 2023-03-21 #5114 @KonradHoeffner
@remram44, @lazear: I hope I have addressed all your comments now.
Please let me know if there is anything more I can do.
The latest article proof has the wrong table layout but as it works locally using the docker container, I hope I can fix this soon.
