### 2023-03-22 #5114 @osorensen
> @osorensen: I have a technical problem with the article proof. Table 1 looks fine when I compile it locally but the layout breaks down in the editorialbot output. Do you perhaps know how I can fix this or who I can contact? Attached is how it is correctly produced with the following command:
>
> ```shell
>   docker run --rm \
>       --volume $PWD:/data \
>       --user $(id -u):$(id -g) \
>       --env JOURNAL=joss \
>       openjournals/paperdraft
> ```
>
> [paper.pdf](https://github.com/openjournals/joss-reviews/files/11029243/paper.pdf)

@openjournals/dev could you please help with this?
