### 2023-03-23 b049b60
Separate benchmark setup from execution function.
