+++ benches/bench.rs
@@ -9,23 +9,29 @@ use sophia::api::term::IriRef;
 use sophia::api::term::SimpleTerm;
 use std::fs::File;
 
-fn bench_query(c: &mut Criterion) {
-    let vincent = "http://dbpedia.org/resource/Vincent_Descombes_Sevoie";
-    let type_ = "http://www.w3.org/1999/02/22-rdf-syntax-ns#type";
-    let person = "http://dbpedia.org/ontology/Person";
+const VINCENT: &str = "http://dbpedia.org/resource/Vincent_Descombes_Sevoie";
+const TYPE: &str = "http://www.w3.org/1999/02/22-rdf-syntax-ns#type";
+const PERSON: &str = "http://dbpedia.org/ontology/Person";
+
+fn load() -> HdtGraph {
     let file = File::open("tests/resources/persondata_en.hdt").expect("error opening file");
     //let file = File::open("tests/resources/lscomplete2015.hdt").expect("error opening file");
     //let file = File::open("tests/resources/snikmeta.hdt").expect("error opening file");
     let hdt = Hdt::new(std::io::BufReader::new(file)).unwrap();
-    let vincent_id = hdt.dict.string_to_id(vincent, &IdKind::Subject);
-    let type_id = hdt.dict.string_to_id(type_, &IdKind::Predicate);
-    let person_id = hdt.dict.string_to_id(person, &IdKind::Object);
-    let graph = HdtGraph::new(hdt);
-    let vincent_term = SimpleTerm::Iri(IriRef::new_unchecked(vincent.into()));
-    let type_term = SimpleTerm::Iri(IriRef::new_unchecked(type_.into()));
-    let person_term = SimpleTerm::Iri(IriRef::new_unchecked(person.into()));
-    let twp = |s, p, o| graph.hdt.triples_with_pattern(s, p, o);
+    HdtGraph::new(hdt)
+}
+
+fn bench_query(c: &mut Criterion) {
+    let graph = load();
     let triples = &graph.hdt.triples;
+    let twp = |s, p, o| graph.hdt.triples_with_pattern(s, p, o);
+
+    let vincent_id = graph.hdt.dict.string_to_id(VINCENT, &IdKind::Subject);
+    let type_id = graph.hdt.dict.string_to_id(TYPE, &IdKind::Predicate);
+    let person_id = graph.hdt.dict.string_to_id(PERSON, &IdKind::Object);
+    let vincent_term = SimpleTerm::Iri(IriRef::new_unchecked(VINCENT.into()));
+    let type_term = SimpleTerm::Iri(IriRef::new_unchecked(TYPE.into()));
+    let person_term = SimpleTerm::Iri(IriRef::new_unchecked(PERSON.into()));
 
     // count to prevent optimizing away function call
     let mut group = c.benchmark_group("??? (all)");
@@ -39,7 +45,7 @@ fn bench_query(c: &mut Criterion) {
     group.bench_function("1.1 (vincent, ?, ?) triple IDs", |b| {
         b.iter(|| SubjectIter::with_pattern(triples, &TripleId::new(vincent_id, 0, 0)).count())
     });
-    group.bench_function("1.2 (vincent, ?, ?) str triples", |b| b.iter(|| twp(Some(vincent), None, None).count()));
+    group.bench_function("1.2 (vincent, ?, ?) str triples", |b| b.iter(|| twp(Some(VINCENT), None, None).count()));
     group.bench_function("1.3 (vincent, ?, ?) Sophia triples", |b| {
         b.iter(|| graph.triples_matching(Some(&vincent_term), Any, Any).count())
     });
@@ -50,7 +56,7 @@ fn bench_query(c: &mut Criterion) {
     group.bench_function("2.1 (?, type, ?) triple IDs", |b| {
[…]
