+++ Cargo.toml
@@ -30,7 +30,11 @@ default = ["sophia"]
 sophia = ["dep:sophia"]
 
 [[bench]]
-name = "bench"
+name = "criterion"
+harness = false
+
+[[bench]]
+name = "iai"
 harness = false
 
 [lib]
@@ -43,3 +47,4 @@ opt-level = 1
 pretty_assertions = "1.3"
 env_logger = { version = "0.10", default-features = false, features = ["auto-color"] }
 criterion = { version = "0.4.0", default-features = false, features = ["cargo_bench_support", "html_reports"] }
+iai = "0.1.1"
