### 2023-03-23 16a2452
Enable triple cache for all triple patterns with two variables (for three it is already enabled).
