### 2023-04-05 #5114 @osorensen
👋 @remram44, @lazear, could you please check if the latest updates resolve the points you raised, and if so, please update your review checklists?

@KonradHoeffner, I'm working on the formatting of Table 1, and will get back to you as soon as it's resolved.
