### 2023-04-05 #5114 @tarleb
Tables can be tricky sometimes. A (somewhat) counter-intuitive behavior of the system is that tables with at least one line that's longer than 72 characters get the relative column information from the Markdown source, while tables with short lines will get "auto-tuned". The sources contained tab characters, which probably made the table wider than it might have appeared.

Here's a re-formatted table that renders better.

``` md
| Library         | Memory in MB | Load Time in ms | Query Time in ms |
|:----------------|-------------:|----------------:|-----------------:|
| hdt_cpp         |      **112** |            1985 |              362 |
| sophia_hdt      |          263 |             930 |              355 |
| hdt_rs          |          264 |         **912** |              315 |
| hdt_java (DelayedString) | 738 |            3170 |          **214** |
| hdt_java (String)        | 785 |            3476 |          **321** |
| sophia_lg       |      **834** |       **11656** |               85 |
| sophia          |         1371 |           15990 |           **20** |
| jena (java)     |         5352 |           40400 |              159 |
| n3js (js)       |        12404 |          100820 |              654 |
| rdflib (python) |        14481 |          182002 |              940 |
| librdf (c)      |           -- |              -- |               -- |
```

Please use the `openjournals/inara` Docker image to compile the PDF, as the `paperdraft` image is outdated. I assume that it is still mentioned in some docs, could you point us to those so we can fix them?

Sorry for the inconvenience.
