### 2023-04-06 6cb131b
Change outdated Docker image in compile script. From paperdraft to openjournals/inara. Fix table layout with inara image.
