### 2023-04-06 #5114 @KonradHoeffner
@tarleb: Thank you for the detailed explanation and the re-formatted table!

**The docs are up to date**, apparently my compile script was still using the outdated version I took from  <https://joss.readthedocs.io/en/latest/submitting.html> a while ago.
