### 2023-04-25 #5114 @remram44
Some remarks for the paper:

* Maybe spell out SPO (subject-predicate-object?)
* Typo for `SPO` vs `???`: both say "all constants", I assume the second one should read "all variables"

I didn't run the entire benchmark suite, but I ran some of the tasks with hdt_rs and librdf and took a look at the methodology, I think it looks solid.
