### 2023-04-25 5d66040
fix ??? pattern description and spell out SPO
