### 2023-04-25 #5114 @KonradHoeffner
@remram44 You are right, thanks for pointing out the mistake! I corrected the description of the "???" pattern and spelled out SPO which indeed stands for subject-predicate-object.
