### 2023-04-25 KonradHoeffner/hdt#31 @osorensen
Editorial comments on JOSS paper
I'm reading through the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/5114) now, and will post comments here.

On line 40, I cannot see that HDT-FoQ has been defined yet, so please introduce it here. I notice that it's introduced in the caption to Figure 2, but please do it in the main text also.
