### 2023-04-25 #5114 @osorensen
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.2139/ssrn.3198999 is OK
> - 10.1007/978-3-642-30284-8_36 is OK
> - 10.3233/sw-150172 is OK
> - 10.1007/978-3-642-41338-4_18 is OK
> - 10.1109/TKDE.2020.2966609 is OK
>
> MISSING DOIs
>
> - 10.1109/dcc.2015.16 may be a valid DOI for title: Serializing RDF in compressed space
> - 10.3233/jifs-179888 may be a valid DOI for title: iHDT++: improving HDT for SPARQL triple pattern resolution
> - 10.3233/sw-140134 may be a valid DOI for title: DBpedia – A large-scale, multilingual knowledge base extracted from wikipedia
>
> INVALID DOIs
>
> - None
> ```

@KonradHoeffner, please fix the three missing DOIs mentioned here, by adding the DOI to the BibTeX entry.
