### 2023-04-26 427ea5c
add DOIs and extend bibtex entries
