+++ paper.bib
@@ -6,7 +6,10 @@
   number={2},
   pages={76--83},
   year={2021},
-  publisher={ACM New York, NY, USA}
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  issn = {0001-0782},
+  doi = {10.1145/3397512},
 }
 
 @article{hdt2013,
@@ -48,7 +51,13 @@ url = {https://www2020devtrack.github.io/site/schedule}
   author={Carroll, Jeremy J and Dickinson, Ian and Dollin, Chris and Reynolds, Dave and Seaborne, Andy and Wilkinson, Kevin},
   booktitle={Proceedings of the 13th international World Wide Web conference on Alternate track papers \& posters},
   pages={74--83},
-  year={2004}
+  isbn = {1581139128},
+  doi={10.1145/1013367.1013381},
+  year={2004},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  location = {New York, NY, USA},
+  series = {WWW Alt. '04}
 }
 
 @misc{rdflib,
@@ -111,6 +120,7 @@ url = {https://www2020devtrack.github.io/site/schedule}
 @InProceedings{readyforaction,
   author="Buil-Aranda, Carlos and Hogan, Aidan and Umbrich, J{\"u}rgen and Vandenbussche, Pierre-Yves",
   noeditor="Alani, Harith and Kagal, Lalana and Fokoue, Achille and Groth, Paul and Biemann, Chris and Parreira, Josiane Xavier and Aroyo, Lora and Noy, Natasha and Welty, Chris and Janowicz, Krzysztof",
+  editor="Alani, Harith and Kagal, Lalana and Fokoue, Achille and Groth, Paul and Biemann, Chris and Parreira, Josiane Xavier and others", 
   title="{SPARQL} Web-Querying Infrastructure: Ready for Action?",
   booktitle="The Semantic Web -- ISWC 2013",
   year="2013",
@@ -124,9 +134,10 @@ url = {https://www2020devtrack.github.io/site/schedule}
 @inproceedings{serializingrdf,
   title={Serializing {RDF} in compressed space},
   author={Hern{\'a}ndez-Illera, Antonio and Mart{\'\i}nez-Prieto, Miguel A and Fern{\'a}ndez, Javier D},
-  booktitle={2015 Data Compression Conference},
+  booktitle={Proceedings of the Data Compression Conference 2015},
   pages={363--372},
   year={2015},
+  doi={10.1109/dcc.2015.16},
   organization={IEEE}
 }
 
@@ -138,17 +149,19 @@ url = {https://www2020devtrack.github.io/site/schedule}
   number={2},
   pages={2249--2261},
   year={2020},
+  doi={10.3233/JIFS-179888},
   publisher={IOS Press}
 }
 
 @article{dbpedia,
-  title={{DB}pedia -- {A} large-scale, multilingual knowledge base extracted from wikipedia},
-  author={Lehmann, Jens and Isele, Robert and Jakob, Max and Jentzsch, Anja and Kontokostas, Dimitris and Mendes, Pablo N and Hellmann, Sebastian and Morsey, Mohamed and Van Kleef, Patrick and Auer, S{\"o}ren and others},
+  title={{DB}pedia -- {A} large-scale, multilingual knowledge base extracted from {W}ikipedia},
+  author={Lehmann, Jens and Isele, Robert and Jakob, Max and Jentzsch, Anja and Kontokostas, Dimitris and Mendes, Pablo N and Hellmann, Sebastian and Morsey, Mohamed and Van Kleef, Patrick and Auer, S{\"o}ren and Bizer, Christian},
[…]
