+++ paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 We present the Rust library hdt-rs (named "hdt" in the context of Rust libraries, such as [on crates.io](https://crates.io/crates/hdt)) for the Header Dictionary Triples (HDT) binary RDF compression format.
 This allows the writing of high-performance Rust applications that load and query HDT datasets using triple patterns.
-Existing Rust applications that use the Sophia library [@sophia] can easily and greatly reduce their RAM usage by using the provided Sophia HDT adapter.
+Existing Rust applications that use the [Sophia library](https://crates.io/crates/sophia) [@sophia] can easily and greatly reduce their RAM usage by using the provided Sophia HDT adapter.
 
 # Preliminaries
 
@@ -36,26 +36,26 @@ Existing Rust applications that use the Sophia library [@sophia] can easily and 
 
 The *Resource Description Framework* (RDF) is a data model that represents information using *triples*, each of which consists of a *subject*, a *predicate*, and an *object*.
 A set of triples is called an *RDF graph*, where the subjects and objects can be visualised as nodes and the predicates as labelled, directed edges.
-A predicate is always an *IRI* (Internationalised Resource Identifier), which is a generalisation of a URI that allows additional characters.
+Predicates are always *IRIs* (Internationalised Resource Identifiers), which are generalisations of a URIs that allow additional characters.
 Subjects and objects can also be *blank nodes* and objects can also be *literals*.
 There are several text-based RDF serialisation formats with different compromises between verbosity, ease of automatic processing, and human readability.
 For example, the N-Triples representation of the fact *"the mayor of Leipzig is Burkhard Jung"* from DBpedia [@dbpedia] is:
 
 ```ntriples
-<http://dbpedia.org/resource/Leipzig> <http://dbpedia.org/ontology/mayor>
-    (no linebreak) <http://dbpedia.org/resource/Burkhard_Jung> .
+<http://dbpedia.org/resource/Leipzig> <http://dbpedia.org/ontology/mayor> \
+    <http://dbpedia.org/resource/Burkhard_Jung> .
 ```
 
 ## Triple Patterns
 
-*Triple patterns* allow matching a subset of a graph.
+*Triple patterns* match a subset of a graph.
 Each part of the pattern is either a constant or a variable, resulting in eight different types. 
-We denote the pattern type with all constants as SPO (*subject-predicate-object*, matching one or zero triples) and the type with all variables as ??? (matching all triples in the graph).
+We denote the pattern type with all constants as SPO (*subject-predicate-object*, matches one or zero triples) and the type with all variables as ??? (matches all triples in the graph).
 The other triple patterns are denoted analogously.
 
 ## Header Dictionary Triples
-While text-based RDF serialisation formats can be read by humans, they are too verbose to be practical on large graphs.
-The serialised size of a graph can be drastically lowered by using the Header Dictionary Triples binary RDF format, which can be loaded into memory in compressed form while still allowing efficient queries.
+While text-based RDF serialisation formats can be read by humans, they are too verbose to be practical for large graphs.
+The serialised size of a graph can be drastically lowered by using the Header Dictionary Triples binary RDF format, which can be loaded into memory in compressed form while still allowing for efficient queries.
 The *header* contains metadata as uncompressed RDF that describes the dataset.
 The *dictionary* stores all the *RDF terms* (IRIs, literals, and blank nodes) in the dataset in compressed form using front-coding [@frontcoding],
 and assigns a unique numerical identifier (ID) to each of them.
@@ -65,9 +65,9 @@ This allows the *triples* component to store the adjacency matrix of the graph u
[…]
