### 2023-04-26 77b3b4c
final optimizations
