### 2023-04-26 ad4a388
Define HDT-FoQ in main text. Part of #31.
