### 2023-04-26 KonradHoeffner/hdt#31 @KonradHoeffner
Defined HDT-FoQ in the main text in https://github.com/KonradHoeffner/hdt/commit/ad4a388322955378f9bde6cdc95b272ca56d3262.
