### 2023-04-26 #5114 @KonradHoeffner
Those three DOIs and a few more others new included with commit 427ea5c78c4a9bb2b5c0aebbddf970bdf4759766.

> @KonradHoeffner, please fix the three missing DOIs mentioned here, by adding the DOI to the BibTeX entry.
