### 2023-04-27 ef4d425
Update sucds from 0.6.0 to 0.7.0 and refactor code accordingly. Release 0.1.1.
