+++ src/triples/object_iter.rs
@@ -1,6 +1,7 @@
 use crate::triples::Id;
 use crate::triples::TripleId;
 use crate::triples::TriplesBitmap;
+use sucds::int_vectors::Access;
 
 // see "Exchange and Consumption of Huge RDF Data" by Martinez et al. 2012
 // https://link.springer.com/chapter/10.1007/978-3-642-30284-8_36
@@ -33,8 +34,8 @@ impl<'a> Iterator for ObjectIter<'a> {
         if self.pos_index > self.max_index {
             return None;
         }
-        let pos_y = self.triples.op_index.sequence.get(self.pos_index);
-        let y = self.triples.wavelet_y.get(pos_y) as Id;
+        let pos_y = self.triples.op_index.sequence.access(self.pos_index).unwrap();
+        let y = self.triples.wavelet_y.access(pos_y).unwrap() as Id;
         let x = self.triples.bitmap_y.dict.rank(pos_y as u64, true) as Id + 1;
         self.pos_index += 1;
         Some(TripleId::new(x, y, self.o))
