### 2023-04-27 #5114 @KonradHoeffner
@osorensen sure!

- [x] created new release 0.1.1 <https://github.com/KonradHoeffner/hdt/releases/tag/0.1.0>
- [x] archived hdt 0.1.1 at Zenodo at <https://zenodo.org/record/7871117>
- [x] corrected the metadata
- [x] DOI of archived version: https://zenodo.org/record/7871117
