### 2023-04-27 91db896
Try replacing custom Sequence with sucds 0.7.0 CompactVector. Compiles but fails tests.
