+++ src/containers/sequence.rs
@@ -1,26 +1,16 @@
 use crate::containers::vbyte::read_vbyte;
-use bytesize::ByteSize;
-use std::fmt;
+//use bytesize::ByteSize;
+//use std::fmt;
 use std::io;
 use std::io::BufRead;
 use std::mem::size_of;
 use std::thread;
+use sucds::int_vectors::CompactVector;
+use sucds::Serializable;
 
 const USIZE_BITS: usize = usize::BITS as usize;
 
-/// Integer sequence with a given number of bits, which means numbers may be represented along byte boundaries.
-//#[derive(Clone)]
-pub struct Sequence {
-    /// Number of integers in the sequence.
-    pub entries: usize,
-    /// Number of bits that each integer uses.
-    pub bits_per_entry: usize,
-    /// Data in blocks.
-    pub data: Vec<usize>,
-    /// whether CRC check was successful
-    pub crc_handle: Option<thread::JoinHandle<bool>>,
-}
-
+/*
 impl fmt::Debug for Sequence {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
@@ -32,151 +22,100 @@ impl fmt::Debug for Sequence {
         )
     }
 }
+*/
 
-pub struct SequenceIter<'a> {
-    sequence: &'a Sequence,
-    i: usize,
-}
+/// Read sequence including metadata from HDT data.
+pub fn read_sequence<R: BufRead>(reader: &mut R) -> io::Result<(CompactVector, thread::JoinHandle<bool>)> {
+    use io::Error;
+    use io::ErrorKind::InvalidData;
+    use io::ErrorKind::Other;
 
-impl<'a> Iterator for SequenceIter<'a> {
-    type Item = usize;
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.i >= self.sequence.entries {
-            return None;
-        }
-        let e = self.sequence.get(self.i);
-        self.i += 1;
-        Some(e)
+    // read entry metadata
+    // keep track of history for CRC8
+    let mut history: Vec<u8> = Vec::new();
+
+    // read and validate type
+    let mut buffer = [0_u8];
+    reader.read_exact(&mut buffer)?;
+    history.extend_from_slice(&buffer);
+    if buffer[0] != 1 {
+        return Err(Error::new(InvalidData, "Invalid LogArray type"));
     }
-}
 
-impl<'a> IntoIterator for &'a Sequence {
-    type Item = usize;
-    type IntoIter = SequenceIter<'a>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        SequenceIter { sequence: self, i: 0 }
+    // read number of bits per entry
+    let mut buffer = [0_u8];
+    reader.read_exact(&mut buffer)?;
+    history.extend_from_slice(&buffer);
+    let bits_per_entry = buffer[0] as usize;
+    if bits_per_entry > USIZE_BITS {
+        return Err(Error::new(InvalidData, "entry size too large (>64 bit)"));
     }
-}
 
-impl Sequence {
-    /// Get the integer at the given index, counting from 0.
-    pub fn get(&self, index: usize) -> usize {
-        let scaled_index = index * self.bits_per_entry;
-        let block_index = scaled_index / USIZE_BITS;
-        let bit_index = scaled_index % USIZE_BITS;
+    // read number of entries
+    let (entries, bytes_read) = read_vbyte(reader)?;
+    history.extend_from_slice(&bytes_read);
 
-        let mut result;
+    // read entry metadata CRC8
+    let mut crc_code = [0_u8];
+    reader.read_exact(&mut crc_code)?;
+    let crc_code = crc_code[0];
 
[…]
